### 2018-10-19 #262 @nldoc
nlrx
**nlrx** enables you to run agent-based models in [NetLogo ](http://ccl.northwestern.edu/netlogo/)from R.
This makes it possible to design reproducible experiments (which can be
zipped with nlrx and shared between persons). Furthermore, it
coerces the spatial data from NetLogo into raster and sf objects.

```
Package: nlrx
Title: Setup, run and analyze NetLogo model simulations from R via XML
Version: 0.1.0
Authors@R: c(person("Jan", "Salecker", email = "jsaleck@gwdg.de",
                     role = c("aut", "cre"), comment = c(ORCID = '0000-0002-9000-4229')),
             person("Marco", "Sciaini", email = "sciaini.marco@gmail.com", role = c("aut"), comment = c(ORCID = '0000-0002-3042-5435')))
Maintainer: Jan Salecker <jsaleck@gwdg.de>
Description: The nlrx package provides tools to setup, run and analyze NetLogo model simulations in R.
    nlrx experiments use a similar structure as NetLogos Behavior Space experiments.
    However, nlrx offers more flexibility and additional tools for running and anlyzing complex simulation designs and sensitivity analyses.
    The user defines all information that is needed in an intuitive framework, using S4 class objects.
    Experiments are submitted from R to NetLogo via XML files that are dynamically written, based on specifications defined by the user.
    By nesting model calls in future environments, large simulation design with many runs can be executed in parallel.
    This also enables simulating NetLogo experiments on remote HPC machines.
    XML,
    lhs,
    sensitivity,
    furrr,
    GenSA,
    genalg,
    raster,
    rstudioapi,
    sf,
    tidyr
```

  - https://github.com/nldoc/nlrx

  - reproducibility, as it is the first package that lets you perform systematic packages with NetLogo from R
  - partly spatial, as the landscapes from NetLogo are retrieved as rasters and the agents as sf objects

  - The audience is the scientific NetLogo community, that performs experiments within
    NetLogo. They now can use nlrx to parameterise NetLogo and scale their experiments up.
    With **nlrx** it is the first also possible to run NetLogo experiments on computing clusters.

  - There is **RNetLogo** on CRAN, this package however is based on **rJava** (which is most often an obstacle) and has not as much auxiliary functions to run NetLogo and embed results in R objects.

  - Robin Lovelace (@Robinlovelace i), as he has experience with spatial simulations and NetLogo
