### 2018-10-19 #262 @sckott
@nldoc I completely agree the lack of rJava dependency is a big plus. . But can you comment more on the benefit of your pkg over `RNetLogo` with respect to features/functionality.
