### 2018-10-19 #262 @sckott
There's also https://cran.rstudio.com/web/packages/NetLogoR/ - not sure if you were aware or not?  Can you also tell us benefit of your pkg over `NetLogoR` as well please
