### 2018-10-20 #262 @nldoc
>
>
> @nldoc I completely agree the lack of rJava dependency is a big plus. . But can you comment more on the benefit of your pkg over `RNetLogo` with respect to features/functionality.

The main purpose of nlrx is running a NetLogo model and collecting output data. Everything is designed to make this as simple and reproducible as possible.

Reproducability
* nlrx uses nested S4 classes to store all the information that is needed to run NetLogo model simulations.
* After running simulations, the model output is attached to the S4 class object, thus the complete experiment including all experiment definitions, the parameter input matrix, runtime, etc. is now stored in one single S4 class object that can be easily archived and shared with other users.
* For complete reproducability the NetLogo model files are needed. Thus, nlrx provides a function to generate a zip file from the model folder, including a `rds` file containing the S4 class object that holds the information on the experiment. This zip file could be given to anyone to reproduce nlrx NetLogo simulations.
* When using RNetLogo, everything has to be written from scratch and there is no easy way to share and reproduce complex simulation experiments.

Setting up experiments
* Because everything in nlrx is designed on running NetLogo simulations (and in particular Sensitivity Analysis), the setup is very intuitive and requires a minimum amount of code to be written.
* `RNetLogo` may be more general in its usage than nlrx because it allows controlling interactive NetLogo sessions from R, however to set up large sensitivity analysis experiments a lot of R code needs to be written.
* The nlrx design is in many parts adapted from the NetLogo built-in tool behavior space, which allows easy understanding of experiment setup for users that are already familiar with NetLogo

Auxiliary functions
* The nlrx package provides a collection of simulation design helper functions. These can be used to create parameter input matrices for different simulation methods, based on the defined model variables.
* For example, the `simdesign_ff()` creates a full factorial parametrer matrix. `simdesign_morris()` creates a morris screening parameter matrix (there are many more).
* After running a sensitivity analysis simdesign created with one of these helper functions (e.g. `simdesign_morris()`), nlrx provides functions to calculate sensitivity indices for each parameter and output
* RNetLogo does not provide any auxiliary functions to setup simulations, generate input parameter matrices or analyze model output
* Additionally, nlrx provides auxiliary functions that make it easy to convert spatial output from NetLogo models to spatial R formats (patches are stored as raster objects and turtles are stored as sf point objects). RNetLogo only provides functions to collect spatial data from NetLogo as data frames.

Summary
The main purpose of the `nlrx` package was not to replace `RNetLogo`. Our motivation was to develop a package that makes running sensitivity analysis with R and NetLogo easier and reproducible. While `RNetLogo` may be more general in its usage, `nlrx` has many benefits when it comes to its main task: running a NetLogo model many times, with different parameter combinations and collecting model output. For this task `nlrx` is better suited as `RNetLogo` because of easier setup (with less code), enhanced reproducability and easy exchange with other users, built-in auxiliary functions to create parameter matrices and calculate sensitivity indices, and increased stability by removing `rJava` dependency.
