### 2018-10-20 #262 @nldoc
>
>
> There's also https://cran.rstudio.com/web/packages/NetLogoR/ - not sure if you were aware or not? Can you also tell us benefit of your pkg over `NetLogoR` as well please

The `NetLogoR` package follows a different approach because it re-implements the NetLogo syntax in R. While this package can be used to rebuild NetLogo models in R, it can not be used to control already existing NetLogo models, which is the main purpose of the `nlrx` package.
