### 2018-10-22 #262 @marcosci
> @nldoc I completely agree the lack of rJava dependency is a big plus... But can you comment more on the benefit of your pkg over `RNetLogo` with respect to features/functionality?

What Jan (@nldoc ) did not mention is the way the Java virtual machine is handled - **nlrx** opens a new one for each NetLogo model run, whereas **RNetLogo** keeps a single one open until you restart the R session. This can be quite cumbersome, in terms of memory, and should be solved with the way **nlrx** handles this.
