### 2018-10-24 9c0538d
small updates
