+++ R/class_init.R
@@ -1,4 +1,3 @@
-
 .initClasses <- function() {
 
   #' @title S4 class \code{"simdesign"}
@@ -14,22 +13,22 @@
   #' @export
   simdesign <- methods::setClass("simdesign",
 
-                    slots=list(
-                      simmethod="character",
-                      siminput="tbl",
-                      simobject="list",
-                      simseeds="numeric",
-                      simoutput="tbl"
-                    ),
+                                 slots=list(
+                                   simmethod="character",
+                                   siminput="tbl",
+                                   simobject="list",
+                                   simseeds="numeric",
+                                   simoutput="tbl"
+                                 ),
 
-                    prototype=list(
-                      simmethod=NA_character_,
-                      siminput=tibble::tibble(),
-                      simobject=list(),
-                      simseeds=NA_integer_,
-                      simoutput=tibble::tibble()
-                    )
-   )
+                                 prototype=list(
+                                   simmethod=NA_character_,
+                                   siminput=tibble::tibble(),
+                                   simobject=list(),
+                                   simseeds=NA_integer_,
+                                   simoutput=tibble::tibble()
+                                 )
+  )
 
   #' @title S4 class \code{"experiment"}
   #' @description empty
@@ -57,46 +56,46 @@
   #' @export
   experiment <- methods::setClass("experiment",
 
-                    slots=list(
-                      expname="character",
-                      outpath="character",
-                      repetition="numeric",
-                      tickmetrics="character",
-                      idsetup="character",
-                      idgo="character",
-                      idfinal="character",
-                      idrunnum="character",
-                      runtime="numeric",
-                      evalticks="numeric",
-                      stopcond="character",
-                      metrics="character",
-                      metrics.turtles="character",
-                      metrics.patches="character",
-                      metrics.links="character",
-                      variables="list",
-                      constants="list"
-                    ),
+                                  slots=list(
+                                    expname="character",
+                                    outpath="character",
+                                    repetition="numeric",
+                                    tickmetrics="character",
+                                    idsetup="character",
+                                    idgo="character",
+                                    idfinal="character",
+                                    idrunnum="character",
+                                    runtime="numeric",
+                                    evalticks="numeric",
+                                    stopcond="character",
+                                    metrics="character",
+                                    metrics.turtles="character",
+                                    metrics.patches="character",
+                                    metrics.links="character",
+                                    variables="list",
+                                    constants="list"
+                                  ),
 
-                    prototype=list(
-                      expname=NA_character_,
-                      outpath=NA_character_,
-                      repetition=1,
-                      tickmetrics="true",
-                      idsetup="setup",
-                      idgo="go",
-                      idfinal=NA_character_,
-                      idrunnum=NA_character_,
-                      runtime=NA_integer_,
-                      evalticks=NA_integer_,
-                      stopcond=NA_character_,
-                      metrics=NA_character_,
-                      metrics.turtles=NA_character_,
-                      metrics.patches=NA_character_,
-                      metrics.links=NA_character_,
-                      variables=list(),
-                      constants=list()
-                    )
-   )
+                                  prototype=list(
+                                    expname=NA_character_,
+                                    outpath=NA_character_,
+                                    repetition=1,
+                                    tickmetrics="true",
+                                    idsetup="setup",
+                                    idgo="go",
+                                    idfinal=NA_character_,
+                                    idrunnum=NA_character_,
+                                    runtime=NA_integer_,
+                                    evalticks=NA_integer_,
+                                    stopcond=NA_character_,
+                                    metrics=NA_character_,
+                                    metrics.turtles=NA_character_,
+                                    metrics.patches=NA_character_,
[…]
