+++ R/analyze_nl.R
@@ -77,11 +77,14 @@ write_simoutput <- function(nl) {
 #'
 #' @export
 
-analyze_nl <- function(nl, metrics=getexp(nl, "metrics"), funs=dplyr::funs(mean)) {
+analyze_nl <- function(nl, metrics=getexp(nl, "metrics"),
+                       funs=dplyr::funs(mean)) {
 
   ## Check if results have been attached:
   if (purrr::is_empty(getsim(nl, "simoutput"))) {
-    stop("In order to run analyze_nl, output results have to be attached to the simdesign of the nl object first: setsim(nl, \"simoutput\") <- results")
+    stop("In order to run analyze_nl, output results have to be attached to the
+         simdesign of the nl object first: setsim(nl, \"simoutput\")
+         <- results")
   }
 
 
@@ -138,7 +141,8 @@ analyze_ff <- function(nl, metrics, funs) {
 
   ## For lhs we compute mean and sd values of each run/tick combination:
   ffagg <- getsim(nl, "simoutput") %>%
-    dplyr::group_by_at(dplyr::vars("siminputrow", "[step]", names(getsim(nl, "siminput")))) %>%
+    dplyr::group_by_at(dplyr::vars("siminputrow", "[step]",
+                                   names(getsim(nl, "siminput")))) %>%
     dplyr::summarise_at(metrics, funs) %>%
     dplyr::ungroup()
 
@@ -163,7 +167,8 @@ analyze_lhs <- function(nl, metrics, funs) {
 
   ## For lhs we compute mean and sd values of each run/tick combination:
   lhsagg <- getsim(nl, "simoutput") %>%
-    dplyr::group_by_at(dplyr::vars("siminputrow", "[step]", names(getsim(nl, "siminput")))) %>%
+    dplyr::group_by_at(dplyr::vars("siminputrow", "[step]",
+                                   names(getsim(nl, "siminput")))) %>%
     dplyr::summarise_at(getexp(nl, "metrics"), funs) %>%
     dplyr::ungroup()
 
