### 2018-10-24 b51ae34
get rid of 1:ncol
