+++ R/util_runnl.R
@@ -464,7 +464,7 @@ util_read_write_batch <- function(nl) {
 
       # split patches owns into unique elements of a vector
       patches_owns <- toupper(strsplit(patches_string[x,][[1]][[1]][patches_index], " ")[[1]])
-      patches_owns <- as.data.frame(matrix(patches_owns, nrow = 1, byrow = T), stringsAsFactors=FALSE)
+      patches_owns <- as.data.frame(matrix(patches_owns, nrow = 1, byrow = TRUE), stringsAsFactors=FALSE)
       patches_owns <- utils::type.convert(patches_owns)
 
       #patches_owns <- as.data.frame(matrix(patches_owns, nrow = 1), stringsAsFactors=FALSE)
