### 2018-10-25 2e3049a
Update simdesign_helper.R
correct simdesign message typo
