### 2018-10-25 #262 @annakrystalli
---

Hello @nldoc 👋,

Thanks for your package submission! I look forward to learning more about it during the review. I've completed the editors checks and there's a few issues to clear before we proceed.

All tests and checks pass. However:

#####  test coverage

17% test coverage is generally too low to be accepted currently and will likely need more tests. Particular attention is required on the scripts with 0% coverage. Is there are reason these scripts aren't currently being tested?

```r
covr::package_coverage(pkg_dir)
```
```
nlrx Coverage: 17.55%
R/class_init.R: 0.00%
R/download_netlogo.R: 0.00%
R/export_nl.R: 0.00%
R/get_nl_spatial.R: 0.00%
R/import_nl.R: 0.00%
R/run_nl.R: 0.00%
R/simdesign_helper.R: 0.00%
R/util_eval.R: 0.00%
R/util_runnl_dyn.R: 0.00%
R/util_runnl.R: 0.00%
R/util_stuff.R: 0.00%
R/zzz.R: 0.00%
R/class_setget.R: 33.33%
R/analyze_nl.R: 94.02%
R/class_constr.R: 100.00%

```
#####  Good Practice

```r
goodpractice::gp(pkg_dir)
```

**It is good practice to**:

- [ ] not use "Depends" in DESCRIPTION, as it can
    cause name clashes, and poor interaction with other
    packages. Use "Imports" instead.

- [ ] avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows that
    are about 80 characters wide. Try make your lines shorter
    than 80 characters. Have a look at [styler](https://github.com/r-lib/styler) pkg.

    ```
    R/class_init.R:37:1
    R/class_init.R:38:1
    R/class_init.R:39:1
    R/class_init.R:40:1
    R/class_init.R:41:1
    ```

- [ ] avoid calling `setwd()`, it changes the global
    environment. If you need it, consider using `on.exit()` to
    restore the working directory.

    ```
    R/export_nl.R:39:3
    R/export_nl.R:40:11
    R/export_nl.R:48:3
    ```

- [ ] not import packages as a whole, as this can
    cause name clashes between the imported packages. Instead,
    import only the specific functions you need.

- [ ] fix this R CMD check WARNING: LaTeX errors when
    creating PDF version. This typically indicates Rd
    problems.

    ```
    LaTeX errors found: ! Paragraph ended before
    \Rd@code was complete. <to be read again> \par l.100 !
    Paragraph ended before \Rd@code was complete. <to be read
    again> \par l.125
    ```

- [ ] avoid '`T`' and '`F`', as they are just variables
    which are set to the logicals `TRUE` and `FALSE` by
    default, but are not reserved words and hence can be
    overwritten by the user.  Hence, one should always use
    `TRUE` and `FALSE` for the logicals.

    ```
    R/analyze_nl.R:NA:NA
    R/analyze_nl.R:NA:NA
    ```

##### spell check

I've pulled out a few typos that looked real. Might be worth running spell check to satisfy yourself too.

```r
devtools::spell_check(pkg_dir)
```
```
  WORD              FOUND IN
anlyzing          description:3
contatining       simdesign.Rd:20
reproducability   import_nl.Rd:25
```

The biggest issue currently is the test coverage. How long do you think it will take to develop new tests? I can begin looking for reviewers but we'll really need the tests to be complete before they begin and it's best to contact reviewers when the package is ready for review. Let me know if you'd like me to put the review on hold or contact reviewers if you think you can address the issues swiftly.
