+++ R/export_nl.R
@@ -16,11 +16,10 @@
 #'
 #' @examples
 #' \dontrun{
-#'
+#' 
 #' folder <- "/home/user/test"
 #' outfile <- "/home/user/test.zip"
-#' export_nl(nl, folder=folder, outfile=outfile)
-#'
+#' export_nl(nl, folder = folder, outfile = outfile)
 #' }
 #' @aliases export_nl
 #' @rdname export_nl
@@ -29,15 +28,13 @@
 
 
 export_nl <- function(nl,
-                      folder=dirname(getnl(nl, "modelpath")),
-                      outfile)
-{
-
+                      folder = dirname(getnl(nl, "modelpath")),
+                      outfile) {
   mywd <- getwd()
 
   ## Create zip:
   setwd(folder)
-  utils::zip(zipfile=outfile, files = list.files(), extras = "-r")
+  utils::zip(zipfile = outfile, files = list.files(), extras = "-r")
 
   ## Add nl object:
   ## Create a rds file from the nl object
@@ -45,9 +42,8 @@ export_nl <- function(nl,
   nltempfile <- paste0(nltempdir, "/nlobject.rds")
   saveRDS(nl, nltempfile)
   setwd(nltempdir)
-  utils::zip(zipfile=outfile, files="nlobject.rds", flags="-j")
+  utils::zip(zipfile = outfile, files = "nlobject.rds", flags = "-j")
 
   ## Reset wd:
   setwd(mywd)
-
 }
