+++ R/util_runnl.R
@@ -12,7 +12,7 @@
 util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
 
   ### Get the current values from the siminput matrix:
-  simdata_run <- getsim(nl, "siminput")[siminputrow,]
+  simdata_run <- getsim(nl, "siminput")[siminputrow, ]
 
   ### Attach a runnum variable if needed:
   if (!is.na(getexp(nl, "idrunnum"))) {
@@ -23,37 +23,45 @@ util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
 
   ### Create XML object:
   nlXML <- XML::newXMLDoc()
-  experiments <- XML::newXMLNode("experiments", doc=nlXML)
-  experiment <- XML::newXMLNode("experiment", attrs=c(name=getexp(nl, "expname"),
-                                                     repetitions=getexp(nl, "repetition"),
-                                                     runMetricsEveryStep=getexp(nl, "tickmetrics")),
-                               parent=experiments)
+  experiments <- XML::newXMLNode("experiments", doc = nlXML)
+  experiment <- XML::newXMLNode("experiment",
+    attrs = c(
+      name = getexp(nl, "expname"),
+      repetitions = getexp(nl, "repetition"),
+      runMetricsEveryStep = getexp(nl, "tickmetrics")
+    ),
+    parent = experiments
+  )
 
   ## Add Setup, go
-  idsetup <- paste(getexp(nl, "idsetup"), sep="\n", collapse="\n")
-  idgo <- paste(getexp(nl, "idgo"), sep="\n", collapse="\n")
-  XML::addChildren(experiment, XML::newXMLNode("setup", idsetup, parent=experiment))
-  XML::addChildren(experiment, XML::newXMLNode("go", idgo, parent=experiment))
+  idsetup <- paste(getexp(nl, "idsetup"), sep = "\n", collapse = "\n")
+  idgo <- paste(getexp(nl, "idgo"), sep = "\n", collapse = "\n")
+  XML::addChildren(experiment, XML::newXMLNode("setup", idsetup, parent = experiment))
+  XML::addChildren(experiment, XML::newXMLNode("go", idgo, parent = experiment))
 
   ## Add final commands if provided:
   if (!is.na(getexp(nl, "idfinal"))) {
-    idfinal <- paste(getexp(nl, "idfinal"), sep="\n", collapse="\n")
+    idfinal <- paste(getexp(nl, "idfinal"), sep = "\n", collapse = "\n")
     XML::addChildren(experiment, XML::newXMLNode("final",
-                                                 idfinal, parent = experiment))
+      idfinal,
+      parent = experiment
+    ))
   }
 
   ## Add timeLimit:
   runtime <- getexp(nl, "runtime")
   XML::addChildren(experiment, XML::newXMLNode("timeLimit",
-                                               attrs = c(steps = runtime),
-                                               parent = experiment))
+    attrs = c(steps = runtime),
+    parent = experiment
+  ))
 
   ## Add stop condition if provided:
   if (!is.na(getexp(nl, "stopcond"))) {
     stopcond <- paste(getexp(nl, "stopcond"), sep = "\n", collapse = "\n")
     XML::addChildren(experiment, XML::newXMLNode("exitCondition",
-                                                 stopcond,
-                                                 parent = experiment))
+      stopcond,
+      parent = experiment
+    ))
   }
 
   ## Add metrics:
@@ -61,61 +69,81 @@ util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
 
[…]
