### 2018-10-25 #262 @annakrystalli
@nldoc @marcosci Apologies! I only just realised this was a pre-submission enquiry.  So the package has been approved for submission! 🎉 If you would like to go ahead, you already have the editors checks. Just let me know how you would like to proceed and sorry for jumping the gun!
