### 2018-10-25 4e16ede
fix most of annas comments (except for the tests)
