+++ R/util_runnl.R
@@ -1,10 +1,12 @@
-#' Create a temporary behavior space xml file to setup NetLogo via commandline
+#' Create a temporary behavior space xml file to setup NetLogo via command line
 #'
-#' @description Create a temporary behavior space xml file to setup NetLogo via commandline
+#' @description Create a temporary behavior space xml file to setup NetLogo via
+#' command line
 #'
 #' @param nl nl object
 #' @param seed random-seed for NetLogo simulation
-#' @param siminputrow row id of the simulation input tibble of the simdesign within the provided nl object
+#' @param siminputrow row id of the simulation input tibble of the simdesign
+#'  within the provided nl object
 #' @param xmlfile filepath where the xml file is stored
 #' @aliases util_create_sim_XML
 #' @rdname util_create_sim_XML
@@ -16,7 +18,8 @@ util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
 
   ### Attach a runnum variable if needed:
   if (!is.na(getexp(nl, "idrunnum"))) {
-    runnum <- tibble(paste0("\"", getexp(nl, "expname"), "_", seed, "_", siminputrow, "\""))
+    runnum <- tibble(paste0("\"", getexp(nl, "expname"), "_", seed, "_",
+                            siminputrow, "\""))
     names(runnum) <- getexp(nl, "idrunnum")
     simdata_run <- cbind(simdata_run, runnum)
   }
@@ -36,7 +39,8 @@ util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
   ## Add Setup, go
   idsetup <- paste(getexp(nl, "idsetup"), sep = "\n", collapse = "\n")
   idgo <- paste(getexp(nl, "idgo"), sep = "\n", collapse = "\n")
-  XML::addChildren(experiment, XML::newXMLNode("setup", idsetup, parent = experiment))
+  XML::addChildren(experiment, XML::newXMLNode("setup", idsetup,
+                                               parent = experiment))
   XML::addChildren(experiment, XML::newXMLNode("go", idgo, parent = experiment))
 
   ## Add final commands if provided:
@@ -140,25 +144,30 @@ util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
   }
 
   ## Use NetLogo specific prefix:
-  prefix <- "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<!DOCTYPE experiments SYSTEM \"behaviorspace.dtd\">"
+  prefix <- "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<!DOCTYPE experiments
+  SYSTEM \"behaviorspace.dtd\">"
 
   # SAVE XML TO FILE
   cat(XML::saveXML(nlXML, prefix = prefix), file = xmlfile)
 }
 
-#' Setup and execute NetLogo via commandline
+#' Setup and execute NetLogo via command line
 #'
-#' @description Setup and execute NetLogo via commandline
+#' @description Setup and execute NetLogo via command line
 #'
 #' @param nl nl object
 #' @param xmlfile file location of the experiment xml file
 #' @param outfile file location for output results
-#' @param batchfile file location of system specific batch file to call NetLogo via commandline
+#' @param batchfile file location of system specific batch file to call NetLogo
+#' via command line
 #' @aliases util_call_nl
 #' @rdname util_call_nl
 #' @keywords internal
 util_call_nl <- function(nl, xmlfile, outfile, batchfile) {
[…]
