+++ R/export_nl.R
@@ -16,7 +16,7 @@
 #'
 #' @examples
 #' \dontrun{
-#'
+#' 
 #' folder <- "/home/user/test"
 #' outfile <- "/home/user/test.zip"
 #' export_nl(nl, folder = folder, outfile = outfile)
@@ -27,16 +27,14 @@
 #' @export
 
 export_nl <- function(nl,
-                      folder=dirname(getnl(nl, "modelpath")),
-                      outfile)
-{
-
+                      folder = dirname(getnl(nl, "modelpath")),
+                      outfile) {
   mywd <- getwd()
 
   ## Create zip:
   setwd(folder)
   on.exit(setwd(mywd))
-  utils::zip(zipfile=outfile, files = list.files(), extras = "-r")
+  utils::zip(zipfile = outfile, files = list.files(), extras = "-r")
 
   ## Add nl object:
   ## Create a rds file from the nl object
@@ -44,6 +42,5 @@ export_nl <- function(nl,
   nltempfile <- paste0(nltempdir, "/nlobject.rds")
   saveRDS(nl, nltempfile)
   setwd(nltempdir)
-  utils::zip(zipfile=outfile, files="nlobject.rds", flags="-j")
-
+  utils::zip(zipfile = outfile, files = "nlobject.rds", flags = "-j")
 }
