+++ tests/testthat/test_class_construction.R
@@ -3,17 +3,16 @@ testthat::test_that("class objects are created with correct variables", {
   nl <- nl(
     nlversion = "6.0.2",
     nlpath = "C:/Program Files/NetLogo 6.0.2/",
-    modelpath = "C:/Program Files/NetLogo 6.0.2/app/models/
-    Sample Models/Biology/Wolf Sheep Predation.nlogo",
+    modelpath =
+      "C:/Program Files/NetLogo 6.0.2/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
     jvmmem = 1024
   )
 
   testthat::context("Class construction: nl")
   testthat::expect_match(getnl(nl, "nlversion"), "6.0.2")
   testthat::expect_match(getnl(nl, "nlpath"), "C:/Program Files/NetLogo 6.0.2/")
-  testthat::expect_match(getnl(nl, "modelpath"), "C:/Program Files/
-                         NetLogo 6.0.2/app/models/Sample Models/Biology/
-                         Wolf Sheep Predation.nlogo")
+  testthat::expect_match(getnl(nl, "modelpath"),
+"C:/Program Files/NetLogo 6.0.2/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
   testthat::expect_equal(getnl(nl, "jvmmem"), 1024)
   testthat::expect_match(class(getnl(nl, "experiment"))[1], "experiment")
   testthat::expect_match(class(getnl(nl, "simdesign"))[1], "simdesign")
@@ -81,7 +80,8 @@ testthat::test_that("class objects are created with correct variables", {
   testthat::expect_match(names(getexp(nl, "constants"))[7],
                          "show-energy?")
 
-  testthat::expect_match(getexp(nl, "constants")[[1]], "\"sheep-wolves-grass\"")
+  testthat::expect_match(getexp(nl, "constants")[[1]],
+                         "\"sheep-wolves-grass\"")
   testthat::expect_equal(getexp(nl, "constants")[[2]], 30)
   testthat::expect_equal(getexp(nl, "constants")[[3]], 4)
   testthat::expect_equal(getexp(nl, "constants")[[4]], 20)
