+++ man/simdesign.Rd
@@ -4,9 +4,9 @@
 \alias{simdesign}
 \title{Construct a new simdesign object}
 \usage{
-simdesign(simmethod = character(), siminput = tibble(),
-  simobject = list(), simseeds = NA_integer_, simoutput = tibble(),
-  ...)
+simdesign(simmethod = character(), siminput = tibble::tibble(),
+  simobject = list(), simseeds = NA_integer_,
+  simoutput = tibble::tibble(), ...)
 }
 \arguments{
 \item{simmethod}{character string defining the method of the simulation design}
