+++ NAMESPACE
@@ -33,6 +33,4 @@ export(simdesign_sobol)
 export(simdesign_sobol2007)
 export(simdesign_soboljansen)
 export(write_simoutput)
-import(methods)
-import(tibble)
 importFrom(magrittr,"%>%")
