+++ R/util_runnl_dyn.R
@@ -53,7 +53,7 @@ util_run_nl_dyn_GenSA_fn <- function(param, nl, evalcrit, seed, cleanup) {
   # Generate a parameterset:
   names(param) <- names(getexp(nl, "variables"))
 
-  gensa_param <- tibble::as.tibble(cbind(as.tibble(t(param)),
+  gensa_param <- tibble::as.tibble(cbind(tibble::as.tibble(t(param)),
     getexp(nl, "constants"),
     stringsAsFactors = FALSE
   ))
@@ -136,7 +136,7 @@ util_run_nl_dyn_GenAlg_fn <- function(param, nl, evalcrit, seed, cleanup) {
   # Generate a parameterset:
   names(param) <- names(getexp(nl, "variables"))
 
-  gensa_param <- tibble::as.tibble(cbind(as.tibble(t(param)),
+  gensa_param <- tibble::as.tibble(cbind(tibble::as.tibble(t(param)),
     getexp(nl, "constants"),
     stringsAsFactors = FALSE
   ))
