+++ man/nl.Rd
@@ -6,7 +6,8 @@
 \usage{
 nl(nlversion = "6.0.2", nlpath = character(),
   modelpath = character(), jvmmem = 1024,
-  experiment = new("experiment"), simdesign = new("simdesign"), ...)
+  experiment = methods::new("experiment"),
+  simdesign = methods::new("simdesign"), ...)
 }
 \arguments{
 \item{nlversion}{A character string defining the NetLogo version that is used}
