### 2018-10-25 #262 @nldoc
Dear @annakrystalli and @sckott, thank you very much for your encouraging response and first feedback.
We are very happy about the chance to further improve our package. We already started fixing most of the issues you mentioned:

>
> ### Editor checks:
>
>     * [x]  **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
>
>     * [ ]  **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
>
>     * [x]  **License:** The package has a CRAN or OSI accepted license
>
>     * [x]  **Repository:** The repository link resolves correctly
>
>     * [ ]  **Archive** (JOSS only, may be post-review): The repository DOI resolves correctly
>
>     * [ ]  **Version** (JOSS only, may be post-review): Does the release version given match the GitHub release (v1.0.0)?
>
>
> #### Editor comments
>
> Hello @nldoc 👋,
>
> Thanks for your package submission! I look forward to learning more about it during the review. I've completed the editors checks and there's a few issues to clear before we proceed.
>
> All tests and checks pass. However:
> ##### test coverage
>
> 17% test coverage is generally too low to be accepted currently and will likely need more tests. Particular attention is required on the scripts with 0% coverage. Is there are reason these scripts aren't currently being tested?
>
> ```r
> covr::package_coverage(pkg_dir)
> ```
>
> ```
> nlrx Coverage: 17.55%
> R/class_init.R: 0.00%
> R/download_netlogo.R: 0.00%
> R/export_nl.R: 0.00%
> R/get_nl_spatial.R: 0.00%
> R/import_nl.R: 0.00%
> R/run_nl.R: 0.00%
> R/simdesign_helper.R: 0.00%
> R/util_eval.R: 0.00%
> R/util_runnl_dyn.R: 0.00%
> R/util_runnl.R: 0.00%
> R/util_stuff.R: 0.00%
> R/zzz.R: 0.00%
> R/class_setget.R: 33.33%
> R/analyze_nl.R: 94.02%
> R/class_constr.R: 100.00%
> ```

We will develop tests for all functions (where possible) within the next two weeks. Unfortunately, some functions (especially those that call java via command line) can not be tested without providing an executable NetLogo version within the repository. However, including NetLogo in the package repository would need way too much space.

> ##### Good Practice
>
> ```r
> goodpractice::gp(pkg_dir)
> ```
>
> **It is good practice to**:
>
>     * [ ]  not use "Depends" in DESCRIPTION, as it can
>       cause name clashes, and poor interaction with other
>       packages. Use "Imports" instead.
>
>     * [ ]  avoid long code lines, it is bad for
>       readability. Also, many people prefer editor windows that
>       are about 80 characters wide. Try make your lines shorter
>       than 80 characters. Have a look at [styler](https://github.com/r-lib/styler) pkg.
>       ```
>       R/class_init.R:37:1
>       R/class_init.R:38:1
>       R/class_init.R:39:1
>       R/class_init.R:40:1
>       R/class_init.R:41:1
>       ... and 98 more lines
>       ```
>
>     * [ ]  avoid calling `setwd()`, it changes the global
>       environment. If you need it, consider using `on.exit()` to
>       restore the working directory.
>       ```
>       R/export_nl.R:39:3
>       R/export_nl.R:40:11
>       R/export_nl.R:48:3
>       ```
>
>     * [ ]  not import packages as a whole, as this can
>       cause name clashes between the imported packages. Instead,
>       import only the specific functions you need.
>
>     * [ ]  fix this R CMD check WARNING: LaTeX errors when
>       creating PDF version. This typically indicates Rd
>       problems.
>       ```
>       LaTeX errors found: ! Paragraph ended before
>       \Rd@code was complete. <to be read again> \par l.100 !
>       Paragraph ended before \Rd@code was complete. <to be read
>       again> \par l.125
>       ```
>
>     * [ ]  avoid '`T`' and '`F`', as they are just variables
>       which are set to the logicals `TRUE` and `FALSE` by
>       default, but are not reserved words and hence can be
>       overwritten by the user.  Hence, one should always use
>       `TRUE` and `FALSE` for the logicals.
>       ```
>       R/analyze_nl.R:NA:NA
>       R/analyze_nl.R:NA:NA
>       ```

We solved most of these issues with our latest updates. The only issue that persists is the use of setwd() within the export_nl function. The function creates a zip file from a defined directory. If the zip file is created without setting the defined directory as working directory, the complete folder directory (back to the current working directory) is stored within the zip file. We did not find any other workaround beside setting the working directory via setwd(). If there are other solutions to create a proper zip file, we are happy to remove the setwd command. However, we followed the suggestion and added the on.exit() command to reset the working directory to the original directory.

>
> ##### spell check
>
> I've pulled out a few typos that looked real. Might be worth running spell check to satisfy yourself too.
>
> ```r
> devtools::spell_check(pkg_dir)
> ```
>
> ```
>   WORD              FOUND IN
> anlyzing          description:3
> contatining       simdesign.Rd:20
> reproducability   import_nl.Rd:25
> ```
>
> The biggest issue currently is the test coverage. How long do you think it will take to develop new tests? I can begin looking for reviewers but we'll really need the tests to be complete before they begin and it's best to contact reviewers when the package is ready for review. Let me know if you'd like me to put the review on hold or contact reviewers if you think you can address the issues swiftly.

As already mentioned, we will develop the remaining tests within the next two weeks. If there are no more issues than the ones stated above the package will be ready for review in two weeks at latest. Thus, we think it makes sense to already contact the reviewers.
