### 2018-10-25 #262 @marcosci
Thanks @annakrystalli for your efforts already :tada:

Regarding the tests:

NetLogo is ~200 Mb big, so, unfortunately, we can not deliver it with the package. This also means that we can not properly test if running NetLogo itself works.

Are you aware of a possibility to do that only on Travis (it could be skipped on CRAN then)?
