### 2018-10-25 #262 @annakrystalli
Thanks for your comments @noamross!

@nldoc, re: **`setwd()` in `export_nl`**

Have you tried supplying the `path` argument to `list.files()` and setting `full.names = TRUE`? This shoulh give you the full paths to the files you want to zip and not require you to set (or reset) the working directory.

eg I think you could do:

```r
 utils::zip(zipfile=outfile, files = list.files(folder, full.names = TRUE), extras = "-r")
```

I also notice you are using `paste0()` to concatenate file paths. It's better to use `file.path()` instead.

eg you could do

```r
nltempfile <- file.path(nltempdir, "nlobject.rds")
```
