### 2018-10-29 c4d87a0
linebreak batch fix
