### 2018-10-29 b288ac1
fix download_netlogo
