+++ .travis.yml
@@ -1,7 +1,7 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: r
-sudo: false
+sudo: true
 dist: trusty
 cache: packages
 latex: false
@@ -17,6 +17,7 @@ matrix:
         - brew link --force curl
 
 before_install:
+  - mkdir "/home/travis/netlogo"
   - wget -qO- "https://ccl.northwestern.edu/netlogo/6.0.3/NetLogo-6.0.3-64.tgz" | tar xvz --directory "/home/travis/netlogo"
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
   - sudo apt-get --yes --force-yes update -qq
