+++ tests/testthat/test_run_nl.R
@@ -65,13 +65,13 @@ testthat::test_that("Run nl", {
   seed <- nl@simdesign@simseeds[1]
   siminputrow <- 1
   util_create_sim_XML(nl, seed, siminputrow, xmlfile)
-
+  testthat::expect_true(file.exists(xmlfile))
   xmlfileread <- XML::xmlParse(file = xmlfile)
   xmlfileread <- XML::xmlToList(xmlfileread)
 
   testthat::expect_equal(xmlfileread$experiment$setup, "setup")
   testthat::expect_equal(xmlfileread$experiment$go, "go")
-  testthat::expect_equal(xmlfileread$experiment$timeLimit[["steps"]], "50")
+  testthat::expect_equal(xmlfileread$experiment$timeLimit[["steps"]], "10")
   testthat::expect_equal(xmlfileread$experiment$metric, "count sheep")
   testthat::expect_true(is.character(
     xmlfileread$experiment$enumeratedValueSet$value[["value"]]))
@@ -79,7 +79,7 @@ testthat::test_that("Run nl", {
     xmlfileread$experiment$enumeratedValueSet$.attrs[["variable"]],
     "initial-number-sheep")
   testthat::expect_equal(
-    xmlfileread$experiment$.attrs[["name"]], "nlrx_simple")
+    xmlfileread$experiment$.attrs[["name"]], "nlrx_test")
   testthat::expect_equal(xmlfileread$experiment$.attrs[["repetitions"]], "1")
   testthat::expect_equal(
     xmlfileread$experiment$.attrs[["runMetricsEveryStep"]], "true")
@@ -110,7 +110,10 @@ testthat::test_that("Run nl", {
   testthat::expect_equal(nrow(results), length(nl@experiment@evalticks))
 
   testthat::context("Cleanup:")
-  util_cleanup(nl, nl@experiment@outpath, "all")
+  util_cleanup(nl, outpath, "all")
+  util_cleanup(nl, dirname(xmlfile), pattern = "all")
+  util_cleanup(nl, dirname(batchpath), pattern = ".bat")
+  util_cleanup(nl, dirname(batchpath), pattern = ".sh")
   testthat::expect_false(file.exists(xmlfile))
   testthat::expect_false(file.exists(batchfile))
   testthat::expect_false(file.exists(outfile))
