+++ tests/testthat/test_run_nl.R
@@ -32,9 +32,9 @@ testthat::test_that("Run nl", {
                               tickmetrics = "true",
                               idsetup = "setup",
                               idgo = "go",
-                              idfinal = "reset-ticks",
-                              runtime = 10,
-                              evalticks = 1:10,
+                              idfinal = NA_character_,
+                              runtime = 2,
+                              evalticks = 1:2,
                               metrics = c("count sheep","count wolves"),
                               metrics.patches = c("pxcor", "pycor", "pcolor"),
                               metrics.turtles = c("who", "breed", "pxcor",
