### 2018-10-29 2ccc6da
.. we should maybe also install java to actually use java :P
