+++ .travis.yml
@@ -21,7 +21,7 @@ before_install:
   - wget -qO- "https://ccl.northwestern.edu/netlogo/6.0.3/NetLogo-6.0.3-64.tgz" | tar xvz --directory "/home/travis/netlogo"
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
   - sudo apt-get --yes --force-yes update -qq
-  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev openjdk-8-jdk
+  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev oracle-java8-set-default
   - Rscript -e 'update.packages(ask = FALSE)'
 
 after_success:
