+++ .travis.yml
@@ -17,6 +17,7 @@ matrix:
         - brew link --force curl
 
 before_install:
+  - sudo apt-get update
   - mkdir "/home/travis/netlogo"
   - wget -qO- "https://ccl.northwestern.edu/netlogo/6.0.3/NetLogo-6.0.3-64.tgz" | tar xvz --directory "/home/travis/netlogo"
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
