+++ tests/testthat.R
@@ -1,3 +1,5 @@
+Sys.unsetenv("R_TESTS")
+
 library(testthat)
 library(nlrx)
 
