+++ tests/testthat/test_run_nl.R
@@ -17,7 +17,8 @@ testthat::test_that("Run nl", {
 
   ## Now we check if we can run a simple simulation:
   ## Step1: Create a nl obejct:
-  modelpath <- file.path(nlpath, "app", "models", "Wolf Sheep Predation.nlogo")
+  modelpath <- file.path(nlpath, "app", "models", "Sample Models",
+                         "Biology", "Wolf Sheep Predation.nlogo")
   nl <- nl(nlversion = "6.0.3",
            nlpath = nlpath,
            modelpath = modelpath,
