### 2018-10-29 18b4edd
testfix_js_spatial
