### 2018-10-29 cf63cca
testfix_js_all
