+++ tests/testthat/test_util_runnl.R
@@ -95,13 +95,12 @@ testthat::test_that("util_runnl", {
   testthat::context("Gather results:")
   results <- util_gather_results(nl, outfile, seed, siminputrow)
   testthat::expect_match(class(results)[1], "tbl_df")
-  testthat::expect_equal(nrow(results), 1)
+  testthat::expect_equal(nrow(results), 2)
 
   testthat::context("Cleanup:")
   util_cleanup(nl, outpath, "all")
   util_cleanup(nl, dirname(xmlfile), pattern = "all")
-  util_cleanup(nl, dirname(batchpath), pattern = ".bat")
-  util_cleanup(nl, dirname(batchpath), pattern = ".sh")
+  util_cleanup(nl, dirname(batchfile), pattern = ".sh")
   testthat::expect_false(file.exists(xmlfile))
   testthat::expect_false(file.exists(batchfile))
   testthat::expect_false(file.exists(outfile))
