### 2018-10-29 fc3cb2c
testfix_js_spatial
