+++ tests/testthat/test_util_runnl.R
@@ -98,8 +98,8 @@ testthat::test_that("util_runnl", {
   testthat::expect_equal(nrow(results), 2)
 
   testthat::context("Cleanup:")
-  util_cleanup(nl, outpath, "all")
-  util_cleanup(nl, dirname(xmlfile), pattern = "all")
+  util_cleanup(nl, outpath, ".csv")
+  util_cleanup(nl, dirname(xmlfile), pattern = ".xml")
   util_cleanup(nl, dirname(batchfile), pattern = ".sh")
   testthat::expect_false(file.exists(xmlfile))
   testthat::expect_false(file.exists(batchfile))
