### 2018-10-29 d6aa435
testfix_js_spatial
