+++ tests/testthat/test_get_nl_spatial.R
@@ -75,15 +75,15 @@ testthat::test_that("Get nl spatial", {
                                     turtle_coords = "px", format = "tibble")
   testthat::expect_match(class(results.spatial)[1], "tbl_df")
 
-  testthat::context("Get spatial data with turtles x: raster/sf")
-  results.spatial <- get_nl_spatial(nl, turtles = TRUE, patches=TRUE,
-                                    turtle_coords = "x", format = "spatial")
-  testthat::expect_match(class(results.spatial)[1], "tbl_df")
-
-  testthat::context("Get spatial data with turtles x: tibble")
-  results.spatial <- get_nl_spatial(nl, turtles = TRUE, patches=TRUE,
-                                    turtle_coords = "x", format = "tibble")
-  testthat::expect_match(class(results.spatial)[1], "tbl_df")
+  # testthat::context("Get spatial data with turtles x: raster/sf")
+  # results.spatial <- get_nl_spatial(nl, turtles = TRUE, patches=TRUE,
+  #                                   turtle_coords = "x", format = "spatial")
+  # testthat::expect_match(class(results.spatial)[1], "tbl_df")
+  #
+  # testthat::context("Get spatial data with turtles x: tibble")
+  # results.spatial <- get_nl_spatial(nl, turtles = TRUE, patches=TRUE,
+  #                                   turtle_coords = "x", format = "tibble")
+  # testthat::expect_match(class(results.spatial)[1], "tbl_df")
 
   testthat::context("Get spatial data without turtles: raster")
   results.spatial <- get_nl_spatial(nl, turtles = FALSE, patches=TRUE,
