### 2018-10-30 9988675
testfix_js_spatial
