### 2018-10-30 fe9b201
testfix util eval and stuff
