### 2018-11-04 da18284
test every version if it can be downloaded
