### 2018-11-04 b33e0e0
fix test and exlude nlrx-package.R and zzz.R
