+++ R/zzz.R
@@ -2,7 +2,8 @@
 #   where <- match(paste("package:", pkgname, sep=""), search())
 #   .initClasses(where)
 # }
-
+#nocov start
 .onLoad <- function(libname, pkgname) {
   .initClasses()
 }
+#nocov end
