+++ tests/testthat/test_download_netlogo.R
@@ -6,7 +6,7 @@ testthat::test_that("Download NetLogo", {
   testthat::skip_if(!identical(Sys.getenv("TRAVIS"), "true"))
 
   ## Version 6.0.0
-  nlversion <- "6.0.0"
+  nlversion <- "6.0"
   nlpath <- tempdir()
   download_netlogo(to = nlpath,
                    version = nlversion,
@@ -14,11 +14,11 @@ testthat::test_that("Download NetLogo", {
 
   ### Check that download worked and netlogo jar file exists:
   testthat::expect_true(file.exists(file.path(nlpath,
-                                              "NetLogo 6.0.0",
+                                              "NetLogo 6.0",
                                               "app",
                                               paste0("netlogo-",
                                                      nlversion,
-                                                     ".jar"))))
+                                                     ".0.jar"))))
   ## Version 6.0.1
   nlversion <- "6.0.1"
   nlpath <- tempdir()
