### 2018-11-04 d4020a9
small fixes
