+++ R/nlrx-package.R
@@ -102,7 +102,7 @@
 #' @name nlrx-package
 #' @author Jan Salecker \email{jsaleck@gwdg.de}
 #' @keywords package
-#nocov start
+# nocov start
 "_PACKAGE"
 
 globalVariables(c(
@@ -126,4 +126,4 @@ globalVariables(c(
   "dist.missing",
   "values.missing"
 ))
-#nocov end
+# nocov end
