+++ tests/testthat/test_download_netlogo.R
@@ -5,6 +5,20 @@ testthat::test_that("Download NetLogo", {
   # Run these tests only on TRAVIS:
   testthat::skip_if(!identical(Sys.getenv("TRAVIS"), "true"))
 
+   ## Version 6.0.0
+  nlversion <- "5.3.1"
+  nlpath <- tempdir()
+  download_netlogo(to = nlpath,
+                   version = nlversion,
+                   extract = TRUE)
+
+  ### Check that download worked and netlogo jar file exists:
+  testthat::expect_true(file.exists(file.path(nlpath,
+                                              "NetLogo 5.3.1",
+                                              "app",
+                                              paste0("netlogo-",
+                                                     nlversion,
+                                                     ".jar"))))
   ## Version 6.0.0
   nlversion <- "6.0"
   nlpath <- tempdir()
