+++ tests/testthat/test_get_nl_spatial.R
@@ -62,6 +62,10 @@ testthat::test_that("Get nl spatial", {
 
   results <- run_nl_all(nl)
 
+  testthat::context("Simoutput attached")
+  testthat::expect_error(get_nl_spatial(nl, turtles = TRUE, patches=TRUE,
+                                        turtle_coords = "px", format = "spatial"))
+
   ## Attach results to nl:
   setsim(nl, "simoutput") <- results
 
