### 2018-11-04 7b8e7ba
check for class inits
