### 2018-11-04 e84a8d8
update class tests
