### 2018-11-04 5505137
have to fix get_nl_spatial
