### 2018-11-04 243325e
fix get_nl_spatial
