### 2018-11-04 40f7d2f
test for idfinal, idrunnum and stopcond in util_runnl
