+++ tests/testthat/test_get_nl_spatial.R
@@ -117,13 +117,13 @@ testthat::test_that("Get nl spatial", {
   testthat::expect_false(exists("results.spatial$turtles_x"))
   testthat::expect_false(exists("results.spatial$turtles_y"))
 
-  # testthat::context("Get spatial data without patches: tibble")
-  # results.spatial <- get_nl_spatial(nl, turtles = TRUE, patches=FALSE,
-  #                                   format = "tibble")
-  # testthat::expect_match(class(results.spatial)[1], "tbl_df")
-  # testthat::expect_false(exists("results.spatial$patches_x"))
-  # testthat::expect_false(exists("results.spatial$patches_y"))
-  # testthat::expect_false(is.null(results.spatial$turtles_x))
-  # testthat::expect_false(is.null(results.spatial$turtles_y))
+  testthat::context("Get spatial data without patches: tibble")
+  results.spatial <- get_nl_spatial(nl, turtles = TRUE, patches=FALSE,
+                                    format = "tibble")
+  testthat::expect_match(class(results.spatial)[1], "tbl_df")
+  testthat::expect_false(exists("results.spatial$patches_x"))
+  testthat::expect_false(exists("results.spatial$patches_y"))
+  testthat::expect_false(is.null(results.spatial$turtles_x))
+  testthat::expect_false(is.null(results.spatial$turtles_y))
 
 })
