+++ tests/testthat/test_class_init.R
@@ -7,16 +7,16 @@ testthat::test_that("class objects are initialized", {
  # exp <- methods::new("experiment")
  # sd <- methods::new("simdesign")
 
- # nl <- nl()
- # exp <- experiment()
- # sd <- simdesign()
- #
- # testthat::expect_match(class(nl)[[1]], "nl")
- # testthat::expect_match(class(exp)[[1]], "experiment")
- # testthat::expect_match(class(sd)[[1]], "simdesign")
- #
- # testthat::expect_match(class(nl())[1], "nl")
- # testthat::expect_match(class(experiment())[1], "experiment")
- # testthat::expect_match(class(simdesign())[1], "simdesign")
+ nl <- nl()
+ exp <- experiment()
+ sd <- simdesign()
+
+ testthat::expect_match(class(nl)[[1]], "nl")
+ testthat::expect_match(class(exp)[[1]], "experiment")
+ testthat::expect_match(class(sd)[[1]], "simdesign")
+
+ testthat::expect_match(class(nl())[1], "nl")
+ testthat::expect_match(class(experiment())[1], "experiment")
+ testthat::expect_match(class(simdesign())[1], "simdesign")
 
 })
