+++ .travis.yml
@@ -21,7 +21,7 @@ before_install:
   - Rscript -e 'update.packages(ask = FALSE)'
 
 after_success:
-  - Rscript -e 'covr::codecov()'
+  - Rscript -e 'covr::codecov(token = "3651cae6-2263-4130-a651-75eb7fc02149")'
 
 notifications:
   email:
