### 2018-11-06 eaf739b
fix bits and pieces
