+++ data_raw/create_sample_data.R
@@ -1,36 +1,43 @@
 create_sample_data <- function() {
 
+  ## Set random seed
   set.seed(593472)
+
+  ## Set nl path, model path and output path
+  nlpath <- "C:/Program Files/NetLogo 6.0.3/"
+  modelpath <- "C:/Program Files/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo"
+  outpath <- "C:/out"
+
   # Create nl
-  nl <- setup_sample_data_nl()
+  nl <- setup_sample_data_nl(nlpath, modelpath, outpath)
   nl@simdesign <- simdesign_simple(nl = nl,
                                    nseeds = 1)
   # Store data
   nl_simple <- run_sample_data(nl)
-  devtools::use_data(nl_simple, compress = "gzip")
+  devtools::use_data(nl_simple, compress = "gzip", overwrite = TRUE)
 
 
   # Create nl
-  nl <- setup_sample_data_nl()
+  nl <- setup_sample_data_nl(nlpath, modelpath, outpath)
   nl@simdesign <- simdesign_ff(nl = nl,
                                nseeds = 1)
   # Store data
   nl_ff <- run_sample_data(nl)
-  devtools::use_data(nl_ff, compress = "gzip")
+  devtools::use_data(nl_ff, compress = "gzip", overwrite = TRUE)
 
 
   # Create nl
-  nl <- setup_sample_data_nl()
+  nl <- setup_sample_data_nl(nlpath, modelpath, outpath)
   nl@simdesign <- simdesign_lhs(nl=nl,
                                 samples=100,
                                 nseeds=1,
                                 precision=3)
   # Store data
   nl_lhs <- run_sample_data(nl)
-  devtools::use_data(nl_lhs, compress = "gzip")
+  devtools::use_data(nl_lhs, compress = "gzip", overwrite = TRUE)
 
   # Create nl
-  nl <- setup_sample_data_nl()
+  nl <- setup_sample_data_nl(nlpath, modelpath, outpath)
   nl@simdesign <- simdesign_sobol(nl=nl,
                                   samples=1000,
                                   sobolorder=2,
@@ -40,11 +47,11 @@ create_sample_data <- function() {
                                   precision=3)
   # Store data
   nl_sobol <- run_sample_data(nl)
-  devtools::use_data(nl_sobol, compress = "gzip")
+  devtools::use_data(nl_sobol, compress = "gzip", overwrite = TRUE)
 
 
   # Create nl
-  nl <- setup_sample_data_nl()
+  nl <- setup_sample_data_nl(nlpath, modelpath, outpath)
   nl@simdesign <- simdesign_sobol2007(nl=nl,
                                       samples=1000,
                                       sobolnboot=100,
@@ -53,11 +60,11 @@ create_sample_data <- function() {
                                       precision=3)
   # Store data
   nl_sobol2007 <- run_sample_data(nl)
-  devtools::use_data(nl_sobol2007, compress = "gzip")
+  devtools::use_data(nl_sobol2007, compress = "gzip", overwrite = TRUE)
 
 
   # Create nl
-  nl <- setup_sample_data_nl()
+  nl <- setup_sample_data_nl(nlpath, modelpath, outpath)
   nl@simdesign <- simdesign_soboljansen(nl=nl,
                                         samples=1000,
                                         sobolnboot=100,
@@ -66,11 +73,11 @@ create_sample_data <- function() {
                                         precision=3)
   # Store data
   nl_soboljansen <- run_sample_data(nl)
-  devtools::use_data(nl_soboljansen, compress = "gzip")
[…]
