### 2018-11-07 70f07b8
clean_metrics fix
