### 2018-11-07 e458c39
streamline a tiiiiny bit get_nl_spatial
