+++ R/get_nl_spatial.R
@@ -50,6 +50,7 @@ get_nl_spatial <- function(nl,
                            turtle_coords = "px",
                            format = "spatial") {
 
+
   ## Check if results have been attached:
   if (purrr::is_empty(getsim(nl, "simoutput"))) {
     stop(
@@ -65,6 +66,12 @@ get_nl_spatial <- function(nl,
       id = seq(1, nrow(getsim(nl, "simoutput"))),
       turtle.dat = rep(NA, nrow(getsim(nl, "simoutput")))
     )
+  } else {
+
+    turtles_tib <- getsim(nl, "simoutput") %>%
+      dplyr::select(-metrics.patches) %>%
+      tidyr::unnest(metrics.turtles)
+
   }
 
   ## If no patches shall be returned, create empty tibble
@@ -73,230 +80,98 @@ get_nl_spatial <- function(nl,
       id = seq(1, nrow(getsim(nl, "simoutput"))),
       patch.dat = rep(NA, nrow(getsim(nl, "simoutput")))
     )
-  }
+  } else {
 
-  ## Get spatial patch data
-  if (all(!is.na(getexp(nl, "metrics.patches"))) &&
-    isTRUE(patches)) {
-
-    ## Check if the appropriate reporter for patch coords was used
-    if (!all(any(getexp(nl, "metrics.patches") %in% c("pxcor")) &
-      any(getexp(nl, "metrics.patches") %in% c("pycor")))) {
-      stop(
-        "get_nl_spatial needs pxcor and pycor for creating raster
-           from patches. Please add pxcor and pycor to metrics.patches."
-      )
-    }
-
-    ## grab x coords
-    x_coord_ind <- grepl(
-      c("pxcor"),
-      names(getsim(nl, "simoutput")$metrics.patches[[1]])
-    )
-    x_coord_ind <- which(x_coord_ind == TRUE)
-
-    ## grab y coords
-    y_coord_ind <- grepl(
-      c("pycor"),
-      names(getsim(nl, "simoutput")$metrics.patches[[1]])
-    )
-    y_coord_ind <- which(y_coord_ind == TRUE)
-
-    ## grab patch variables
-    patches_own <-
-      which(seq_len(ncol(getsim(nl, "simoutput")$metrics.patches[[1]])) %in%
-        c(x_coord_ind, y_coord_ind) == FALSE)
-
-    patches_own_names <-
-      names(getsim(nl, "simoutput")$metrics.patches[[1]])[patches_own]
-
-    ## map through every tick and return results
-    patch.dat <- get_patches(nl,
-                             x_coord_ind,
-                             y_coord_ind,
-                             patches_own,
-                             patches_own_names)
-
-    ## streamline return
-    patches_tib <- tibble::enframe(patch.dat, "id", "patches")
-    patches_tib$step <- getsim(nl, "simoutput")$`[step]`
-    patches_tib$siminputrow <- getsim(nl, "simoutput")$siminputrow
-    patches_tib$`random-seed` <-
-      getsim(nl, "simoutput")$`random-seed`
+    patches_tib <- getsim(nl, "simoutput") %>%
+      dplyr::select(-metrics.turtles) %>%
+      tidyr::unnest(metrics.patches)
 
   }
 
-  ## Get spatial turtle data
-  if (all(!is.na(getexp(nl, "metrics.turtles"))) &&
-    isTRUE(turtles)) {
+  if(format == "spatial"){
 
-    ## Check if the appropriate reporter for turtle coords was used
-    if (!all(any(getexp(nl, "metrics.turtles") %in% c("xcor", "pxcor")) &
-      any(getexp(nl, "metrics.turtles") %in% c("ycor", "pycor")))) {
-      stop(
-        "get_nl_spatial needs pxcor/xcor and pycor/ycor for creating sf points
[…]
