### 2018-11-07 04c6b1d
get_nl_spatial fix
