### 2018-11-07 23b12cf
get_nl_spatial fix
