+++ R/get_nl_spatial.R
@@ -49,51 +49,50 @@ get_nl_spatial <- function(nl,
                            patches = TRUE,
                            turtle_coords = "px",
                            format = "spatial") {
-
-
   ## Check if results have been attached:
   if (purrr::is_empty(getsim(nl, "simoutput"))) {
     stop(
       "In order to run get_nl_spatial, output results have to be attached to
-         the simdesign of the nl object first: setsim(nl, \"simoutput\")
-         <- results"
-    )
+      the simdesign of the nl object first: setsim(nl, \"simoutput\")
+      <- results"
+      )
   }
 
-  ## If no turtles shall be returned, create empty tibble
-  if (!isTRUE(turtles)) {
-    turtles_tib <- tibble::tibble(
-      id = seq(1, nrow(getsim(nl, "simoutput"))),
-      turtle.dat = rep(NA, nrow(getsim(nl, "simoutput")))
-    )
-  } else {
 
-    turtles_tib <- getsim(nl, "simoutput") %>%
-      dplyr::select(-metrics.patches) %>%
-      tidyr::unnest(metrics.turtles)
-
-  }
-
-  ## If no patches shall be returned, create empty tibble
-  if (!isTRUE(patches)) {
-    patches_tib <- tibble::tibble(
-      id = seq(1, nrow(getsim(nl, "simoutput"))),
-      patch.dat = rep(NA, nrow(getsim(nl, "simoutput")))
-    )
-  } else {
-
-    patches_tib <- getsim(nl, "simoutput") %>%
-      dplyr::select(-metrics.turtles) %>%
-      tidyr::unnest(metrics.patches) %>%
-      dplyr::rename(patches_x = pxcor,
-                    patches_y = pycor)
-
-  }
-
-  if(format == "spatial"){
-
+  if (format == "tibble")
+  {
+    ## If no turtles shall be returned, create empty tibble
     if (isTRUE(turtles)) {
+      turtles_tib <- getsim(nl, "simoutput") %>%
+        dplyr::select(-metrics.patches) %>%
+        tidyr::unnest(metrics.turtles)
 
+      agentdata <- turtles_tib
+    }
+
+    ## If no patches shall be returned, create empty tibble
+    if (isTRUE(patches)) {
+      patches_tib <- getsim(nl, "simoutput") %>%
+        dplyr::select(-metrics.turtles) %>%
+        tidyr::unnest(metrics.patches) %>%
+        dplyr::rename(patches_x = pxcor,
+                      patches_y = pycor)
+
+      agentdata <- patches_tib
+    }
+
+    if (all(isTRUE(turtles) && isTRUE(patches)))
+    {
+      turtles_tib$group <- "turtles"
+      patches_tib$group <- "patches"
+      agentdata <- dplyr::full_join(patches_tib, turtles_tib)
+    }
+  }
+
+
+  ## Spatial output
+  if (format == "spatial") {
+    if (isTRUE(turtles)) {
       if (turtle_coords == "px") {
         coord_ind <-
           grepl(c("\\bpxcor\\b|\\bpycor\\b"),
@@ -107,17 +106,16 @@ get_nl_spatial <- function(nl,
                 names(getsim(nl, "simoutput")$metrics.turtles[[1]]))
       }
 
-      turtles <- getsim(nl, "simoutput") %>%
+      turtles_dat <- getsim(nl, "simoutput") %>%
         dplyr::select(-metrics.patches) %>%
         dplyr::pull(metrics.turtles)
 
       turtles_tib <- getsim(nl, "simoutput") %>%
         dplyr::select(-metrics.patches) %>%
-        dplyr::mutate(metrics.turtles = purrr::map(seq_along(turtles), function(x){
[…]
