+++ DESCRIPTION
@@ -38,7 +38,8 @@ Imports:
     sf,
     tidyr,
     methods,
-    tibble
+    tibble,
+    stringr
 Depends:
     R (>= 2.10)
 Suggests: 
