### 2018-11-08 cfce9c8
enhanced data compression
