+++ R/simdesign_helper.R
@@ -225,10 +225,16 @@ simdesign_lhs <- function(nl, samples, nseeds, precision) {
                          samples = samples,
                          precision = precision)
 
-  lhs <- tibble::as.tibble(cbind(lhs,
-                                 getexp(nl, "constants"),
-                                 stringsAsFactors=FALSE))
+  ## Bind constants if any:
+  if(length(getexp(nl, "constants")) > 0)
+  {
+    lhs <- cbind(lhs, getexp(nl, "constants"), stringsAsFactors=FALSE)
+  }
 
+  ## Convert to tibble:
+  lhs <- tibble::as.tibble(lhs)
+
+  ## Generate seeds
   seeds <- util_generate_seeds(nseeds = nseeds)
 
   # Add simdesign to nl
@@ -307,10 +313,18 @@ simdesign_sobol <- function(nl,
                            nboot = sobolnboot,
                            conf=sobolconf)
 
-  soX <- tibble::as.tibble(cbind(so$X,
-                                 getexp(nl, "constants"),
-                                 stringsAsFactors=FALSE))
+  ## Export parameter matrix:
+  soX <- so$X
 
+  ## Bind constants if any:
+  if(length(getexp(nl, "constants")) > 0)
+  {
+    soX <- cbind(soX, getexp(nl, "constants"), stringsAsFactors=FALSE)
+  }
+  ## Convert to tibble
+  soX <- tibble::as.tibble(soX)
+
+  ## Generate seeds
   seeds <- util_generate_seeds(nseeds=nseeds)
 
   # Add simdesign to nl
@@ -386,10 +400,19 @@ simdesign_sobol2007 <- function(nl,
                                nboot = sobolnboot,
                                conf=sobolconf)
 
-  soX <- tibble::as.tibble(cbind(so$X,
-                                 getexp(nl, "constants"),
-                                 stringsAsFactors=FALSE))
+  # Export parameter matrix:
+  soX <- so$X
 
+  ## Bind constants if any:
+  if(length(getexp(nl, "constants")) > 0)
+  {
+    soX <- cbind(soX, getexp(nl, "constants"), stringsAsFactors=FALSE)
+  }
+
+  ## Convert to tibble
+  soX <- tibble::as.tibble(soX)
+
+  ## Generate seeds
   seeds <- util_generate_seeds(nseeds=nseeds)
 
   # Add simdesign to nl
@@ -466,10 +489,19 @@ simdesign_soboljansen <- function(nl,
                                  nboot = sobolnboot,
                                  conf=sobolconf)
 
-  soX <- tibble::as.tibble(cbind(so$X,
-                                 getexp(nl, "constants"),
-                                 stringsAsFactors=FALSE))
+  # Export parameter matrix:
+  soX <- so$X
 
+  ## Bind constants if any:
+  if(length(getexp(nl, "constants")) > 0)
+  {
+    soX <- cbind(soX, getexp(nl, "constants"), stringsAsFactors=FALSE)
+  }
+
+  ## Convert to tibble
+  soX <- tibble::as.tibble(soX)
+
+  ## Generate seeds
   seeds <- util_generate_seeds(nseeds=nseeds)
 
   # Add simdesign to nl
@@ -551,10 +583,19 @@ simdesign_morris <- function(nl,
                             bsup = maxs,
                             scale=TRUE)
 
-  moX <- tibble::as.tibble(cbind(tibble::as.tibble(mo$X),
-                                 getexp(nl, "constants"),
-                                 stringsAsFactors=FALSE))
+  # Export parameter matrix:
+  moX <- tibble::as.tibble(mo$X)
 
[…]
