### 2018-11-08 7a5edc3
empty constants list fix2
