### 2018-11-08 af69138
get_nl_spatial fix
