+++ README.md
@@ -1,16 +1,26 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-nlrx <img src="man/figures/logo.png" align="right" width="150" />
-=================================================================
 
-[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/nldoc/nlrx?branch=master&svg=true)](https://ci.appveyor.com/project/nldoc/nlrx) [![Travis build status](https://travis-ci.org/nldoc/nlrx.svg?branch=master)](https://travis-ci.org/nldoc/nlrx) [![Codecov test coverage](https://codecov.io/gh/nldoc/nlrx/branch/master/graph/badge.svg)](https://codecov.io/gh/nldoc/nlrx?branch=master) [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
+# nlrx <img src="man/figures/logo.png" align="right" width="150" />
 
-The nlrx package provides tools to setup NetLogo simulations in R. It uses a similar structure as NetLogos Behavior Space but offers more flexibility and additional tools for running sensitivity analyses.
+[![AppVeyor build
+status](https://ci.appveyor.com/api/projects/status/github/nldoc/nlrx?branch=master&svg=true)](https://ci.appveyor.com/project/nldoc/nlrx)
+[![Travis build
+status](https://travis-ci.org/nldoc/nlrx.svg?branch=master)](https://travis-ci.org/nldoc/nlrx)
+[![Codecov test
+coverage](https://codecov.io/gh/nldoc/nlrx/branch/master/graph/badge.svg)](https://codecov.io/gh/nldoc/nlrx?branch=master)
+[![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
+[![ropensci](https://badges.ropensci.org/262_status.svg)](https://github.com/ropensci/onboarding/issues/262)
 
-Installation
-------------
+The nlrx package provides tools to setup NetLogo simulations in R. It
+uses a similar structure as NetLogos Behavior Space but offers more
+flexibility and additional tools for running sensitivity analyses.
 
-~~You can install the released version of nlrx from~~ ~~[CRAN](https://CRAN.R-project.org) with:~~ ~~install.packages("nlrx")~~
+## Installation
+
+~~You can install the released version of nlrx from~~
+~~[CRAN](https://CRAN.R-project.org) with:~~
+~~install.packages(“nlrx”)~~
 
 And the development version from [GitHub](https://github.com/) with:
 
@@ -19,16 +29,22 @@ And the development version from [GitHub](https://github.com/) with:
 devtools::install_github("nldoc/nlrx")
 ```
 
-Example
--------
+## Example
 
-The nlrx package uses S4 classes to store basic information on NetLogo, the model experiment and the simulation design. Experiment and simulation design class objects are stored within the nl class object. This allows to have a complete simulation setup within one single R object.
+The nlrx package uses S4 classes to store basic information on NetLogo,
+the model experiment and the simulation design. Experiment and
+simulation design class objects are stored within the nl class object.
+This allows to have a complete simulation setup within one single R
+object.
 
-The following steps guide you trough the process on how to setup, run and analyze NetLogo model simulations with nlrx
+The following steps guide you trough the process on how to setup, run
+and analyze NetLogo model simulations with nlrx
 
 #### Step 1: Create a nl object:
 
-The nl object holds all information on the NetLogo version, a path to the NetLogo directory with the defined version, a path to the model file, and the desired memory for the java virtual machine.
[…]
