+++ R/util_runnl.R
@@ -150,10 +150,20 @@ util_create_sim_XML <- function(nl, seed, siminputrow, xmlfile) {
 #' @rdname util_call_nl
 #' @keywords internal
 util_call_nl <- function(nl, xmlfile, outfile, batchfile) {
-  NLcall <- paste0("\"", batchfile, "\"", " --model ", "\"",
-                   getnl(nl, "modelpath"), "\"", " --setup-file ", "\"",
-                   xmlfile, "\"", " --experiment ", getexp(nl, "expname"),
-                   " --table ", "\"", outfile, "\"", " --threads ", 1)
+  if (util_get_os() == "win" | util_get_os() == "unix")
+  {
+    NLcall <- paste0("\"", batchfile, "\"", " --model ", "\"",
+                     getnl(nl, "modelpath"), "\"", " --setup-file ", "\"",
+                     xmlfile, "\"", " --experiment ", getexp(nl, "expname"),
+                     " --table ", "\"", outfile, "\"", " --threads ", 1)
+  }
+  if (util_get_os() == "mac")
+  {
+    NLcall <- paste0("sh \"", batchfile, "\"", " --model ", "\"",
+                     getnl(nl, "modelpath"), "\"", " --setup-file ", "\"",
+                     xmlfile, "\"", " --experiment ", getexp(nl, "expname"),
+                     " --table ", "\"", outfile, "\"", " --threads ", 1)
+  }
   system(NLcall, wait = TRUE)
 }
 
