### 2018-12-18 #262 @annakrystalli
> I’m open to a review of my review🙂
😜❤️

All good, looks like pretty sound feedback to me! If @nldoc or @marcosci require any clarification I'm sure they'll reach out.
