### 2018-12-19 #262 @nldoc
Thank you very much for your detailed reviews @thomasp85 and @marinapapa
Your comments are very helpful for improving the package functionality and documentation.
The issues you mentioned seem straightforward to us. We will immediately start working on improvements and come back to you with a detailed response to your comments.
