+++ README.Rmd
@@ -24,8 +24,26 @@ knitr::opts_chunk$set(
 [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
 [![ropensci](https://badges.ropensci.org/262_status.svg)](https://github.com/ropensci/onboarding/issues/262)
 
-The nlrx package provides tools to setup NetLogo simulations in R.
-It uses a similar structure as NetLogos Behavior Space but offers more flexibility and additional tools for running sensitivity analyses.
+The nlrx package provides tools to setup and execute NetLogo simulations from R.
+NetLogo is a free, open-source and cross-platform modelling environment for simulating natural and social phenomena.
+NetLogo focusses on implementation of agent-based and spatially explicit simulation models, although system dynamics models are supported as well.
+NetLogo is developed and maintained at the Center for Connected Learning and Computer-Based Modeling, Northwestern University, Evanston, IL.
+More details on NetLogo itself are available online: [NetLogo online documentation](https://ccl.northwestern.edu/netlogo/docs/)
+
+NetLogo comes with the built-in experiment tool [Behavior Space](https://ccl.northwestern.edu/netlogo/docs/behaviorspace.html) that allows to setup and execute model simulations with different settings and parameter variations and to collect model output. This experiment tool can be executed via command line in combination with an XML file that contains the experiment specifications, such as runtime, output measurements, stop conditions, and more. One limitation of Behavior Space is, that it only supports full-factorial parameter designs, which may not be appropriate for complex model analyses. Furthermore, Behavior Space experiment specifications are stored within the NetLogo file and are not easily accessible from R for further reproducible analysis.
+
+The nlrx package utilizes the commandline functionality of Behavior Space to execute NetLogo simulations directly from R.
+Instead of defining experiments within NetLogo Behavior Space, experiments are defined in R using the class objects of the nlrx package.
+These class objects hold all the information that is needed to run these experiments remotely from R, such as path to NetLogo installation folder, path to the model file and the experiment specifications itself. nlrx provides useful helper functions to generate parameter input matrices from parameter range definitions that cover a wide range of parameter exploration approaches.
+
+The nlrx package uses class objects to store all relevant information on simulation experiments, including the output of the model simulations. This allows to easily store and share simulation results including all experiment specifications.
+
+In summary, the nlrx package uses a similar structure as NetLogos Behavior Space but offers more flexibility and additional tools for running complex model analyses directly from R.
+
+## Prerequirements
+
+In order to use the nlrx package, NetLogo needs to be installed on the system that is used to execute model simulations (local/remote). For remote execution, NetLogo needs to be installed on remote machines as well. Because NetLogo is executed in a Java virtual machine, Java needs to be installed on the local/remote system as well. We recommend the [Oracle Java SE Development Kit 8](https://www.oracle.com/technetwork/java/javase/downloads/jdk8-downloads-2133151.html)
+While the nlrx package might work without setting the Java system path explicitly, we recommend to make sure that JAVA_HOME points to the correct Java installation of the system.
 
 ## Installation
 
@@ -40,33 +58,63 @@ And the development version from [GitHub](https://github.com/) with:
 devtools::install_github("nldoc/nlrx")
 ```
 
-## Example
+## Get started
 
-The nlrx package uses S4 classes to store basic information on NetLogo, the model experiment and the simulation design.
-Experiment and simulation design class objects are stored within the nl class object.
-This allows to have a complete simulation setup within one single R object.
+The main focus of nlrx is to define and execute NetLogo model simulation experiments from R.
+All information that is needed to run NetLogo simulations remotely is stored within a nl class object.
[…]
