+++ vignettes/simdesign-examples.Rmd
@@ -17,22 +17,41 @@ knitr::opts_chunk$set(
 ```
 
 
-## Further Wolf Sheep examples for all implemented simdesigns
+## Simdesign examples
+
+nlrx provides a bunch of different simdesigns that generate parameter input matrices from valid experiment variable definitions.
+These simdesigns represent different techniques of parameter space exploration.
+This is not a technical guide on parameter space exploration but rather a collection of code examples on how to setup these different approaches with nlrx.
+Details on the sensitivity analysis methods can be found in the documentation of the [sensitivity package](https://cran.r-project.org/web/packages/sensitivity/index.html).
+Details on the optimization methods can be found in the documentation of the [genalg package](https://cran.r-project.org/web/packages/genalg/index.html) and the [GenSA package](https://cran.r-project.org/web/packages/GenSA/index.html).
+
+### Simdesign examples for the Wolf Sheep Predation model
 
 The following section provides valid experiment setups for all supported simdesigns using the Wolf Sheep Model from the NetLogo models library.
 
 First we set up a nl object. We use this nl object for all simdesigns:
 
 ```{r eval=FALSE}
+library(nlrx)
+# Windows default NetLogo installation path (adjust to your needs!):
+netlogopath <- file.path("C:/Program Files/NetLogo 6.0.3")
+modelpath <- file.path(netlogopath, "app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
+outpath <- file.path("C:/out")
+# Unix default NetLogo installation path (adjust to your needs!):
+netlogopath <- file.path("/home/NetLogo 6.0.3")
+modelpath <- file.path(netlogopath, "app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
+outpath <- file.path("/home/out")
+
 nl <- nl(nlversion = "6.0.3",
-         nlpath = "C:/Program Files/NetLogo 6.0.3/",
-         modelpath = "C:/Program Files/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
+         nlpath = netlogopath,
+         modelpath = modelpath,
          jvmmem = 1024)
 ```
 
-##### simdesign_simple
+#### Simple simulation using constants (simdesign_simple)
 
-To setup a simple simdesign, no variables have to be defined. The simple simdesign only uses defined constants and reports a parameter matrix with only one parameterization.
+The simple simdesign only uses defined constants and reports a parameter matrix with only one parameterization.
+To setup a simple simdesign, no variables have to be defined. 
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
@@ -61,9 +80,9 @@ nl@simdesign <- simdesign_simple(nl=nl,
                                  nseeds=3)
 ```
 
-##### simdesign_distinct
+#### Distinct parameter combinations (simdesign_distinct)
 
-To setup a distinct simdesign, vectors of values need to be defined for each variable. These vectors must have the same number of elemtents across all variables.
+The distinct simdesign can be used to run distinct parameter combinations. To setup a distinct simdesign, vectors of values need to be defined for each variable. These vectors must have the same number of elemtents across all variables. The first simulation run consist of all 1st elements of these variable vectors; the second run uses all 2nd values, and so on.
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
@@ -92,9 +111,10 @@ nl@simdesign <- simdesign_distinct(nl=nl,
 ```
 
 
-##### simdesign_ff
+#### Full-factorial simulation (simdesign_ff)
 
-To setup a full-factorial simdesign, vectors of values need to be defined for each variables. Alternatively, a sequence can be defined by setting min, max and step. However, if both (values and min, max, step) are defined, the values vector is prioritized.
[…]
