### 2018-12-19 eb937a5
remove don't run where necessary
