+++ R/class_setget.R
@@ -5,9 +5,15 @@
 #' @param var valid nl variable string
 #' @param value valid value for the specified variable
 #' @examples
-#' \dontrun{
+#' # Example for Wolf Sheep Predation model from NetLogo models library:
+#' nl <- nl(
+#' nlpath = "/home/user/NetLogo 6.0.3/",
+#' modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
+#' jvmmem = 1024)
+#'
+#' # set NetLogo version
 #' setnl(nl, "nlversion") <- "6.0.3"
-#' }
+#'
 #' @aliases setnl
 #' @rdname setnl
 #'
@@ -24,9 +30,16 @@
 #' @param nl nl object
 #' @param var valid nl variable string
 #' @examples
-#' \dontrun{
+#'
+#' # Example for Wolf Sheep Predation model from NetLogo models library:
+#' nl <- nl(nlversion = "6.0.3",
+#' nlpath = "/home/user/NetLogo 6.0.3/",
+#' modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
+#' jvmmem = 1024)
+#'
+#' # get NetLogo version
 #' getnl(nl, "nlversion")
-#' }
+#'
 #' @aliases getnl
 #' @rdname getnl
 #'
@@ -44,9 +57,16 @@ getnl <- function(nl, var) {
 #' @param var valid experiment variable string
 #' @param value valid value for the specified variable
 #' @examples
-#' \dontrun{
+#'
+#' # Example for Wolf Sheep Predation model from NetLogo models library:
+#' nl <- nl(nlversion = "6.0.3",
+#' nlpath = "/home/user/NetLogo 6.0.3/",
+#' modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
+#' jvmmem = 1024)
+#'
+#' # Set experiment name
 #' setexp(nl, "expname") <- "experimentName"
-#' }
+#'
 #' @aliases setexp
 #' @rdname setexp
 #'
@@ -63,9 +83,18 @@ getnl <- function(nl, var) {
 #' @param nl nl object
 #' @param var valid experiment variable string
 #' @examples
-#' \dontrun{
-#' getexp(nl, "expname")
-#' }
+#' # Example for Wolf Sheep Predation model from NetLogo models library:
+#' nl <- nl(nlversion = "6.0.3",
+#' nlpath = "/home/user/NetLogo 6.0.3/",
+#' modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
+#' jvmmem = 1024)
+#'
+#' # Set experiment name
+#' setexp(nl, "expname") <- "experimentName"
+#'
+#' # Get experiment name
+#' getexp(nl, "experiment")
+#'
 #' @aliases getexp
 #' @rdname getexp
 #'
@@ -83,9 +112,15 @@ getexp <- function(nl, var) {
 #' @param var valid simdesign variable string
 #' @param value valid value for the specified variable
 #' @examples
-#' \dontrun{
+#' # Example for Wolf Sheep Predation model from NetLogo models library:
+#' nl <- nl(nlversion = "6.0.3",
+#' nlpath = "/home/user/NetLogo 6.0.3/",
+#' modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
[…]
