+++ R/class_constr.R
@@ -21,14 +21,11 @@
 #'
 #'
 #' @examples
-#' \dontrun{
 #' # Example for Wolf Sheep Predation model from NetLogo models library:
 #' nl <- nl(nlversion = "6.0.3",
-#' nlpath = "C:/Program Files/NetLogo 6.0.3/",
-#' modelpath = "C:/Program Files/NetLogo 6.0.3/app/models/
-#' Sample Models/Biology/Wolf Sheep Predation.nlogo",
+#' nlpath = "/home/user/NetLogo 6.0.3/",
+#' modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
 #' jvmmem = 1024)
-#' }
 #'
 #' @name nl
 #' @rdname nl
