+++ R/util_runnl.R
@@ -324,10 +324,6 @@ util_gather_results <- function(nl, outfile, seed, siminputrow) {
 }
 # nocov end
 
-
-
-
-
 #' Write a modified batchfile that executes NetLogo
 #'
 #' @description Write a modified batchfile that executes NetLogo
@@ -495,156 +491,3 @@ util_read_write_batch <- function(nl) {
 
   return(batchpath_temp)
 }
-
-
-#
-#
-#
-#
-# ## Clean patch metrics
-# .util_clean_metrics_patches <- function(NLtable, nl) {
-#   patches_string <- NLtable[, grepl(c("metrics.patches"), names(NLtable))]
-#
-#   patches_owns <- purrr::map(seq_len(nrow(patches_string)), function(x) {
-#     patches_breed <- regmatches(patches_string[x, ][[1]],
-#                                 gregexpr("(?<=\\{).*?(?=\\})",
-#                                          patches_string[x, ][[1]], perl = TRUE))
-#
-#     patches_string[x, ][[1]] <- substring(patches_string[x, ][[1]], 2)
-#     patches_string[x, ][[1]] <-
-#       substring(patches_string[x, ][[1]],
-#                 1,
-#                 max(nchar(patches_string[x, ][[1]])) - 1)
-#
-#     patches_string[x, ][[1]] <- gsub("(?<=\\{).*?(?=\\})",
-#                                      "\\1",
-#                                      patches_string[x, ][[1]],
-#                                      perl = TRUE)
-#     patches_string[x, ][[1]] <- gsub("\\{\\}",
-#                                      "\\1",
-#                                      patches_string[x, ][[1]],
-#                                      perl = TRUE)
-#     patches_string[x, ][[1]] <- regmatches(patches_string[x, ][[1]],
-#                                            gregexpr("(?<=\\[).*?(?=\\])",
-#                                                     patches_string[x, ][[1]],
-#                                                     perl = TRUE))
-#
-#     patches_owns <-
-#       suppressWarnings(purrr::map_dfr(seq_along(patches_string[x, ][[1]][[1]]),
-#                                       function(patches_index) {
-#
-#       # split patches owns into unique elements of a vector
-#       patches_owns <-
-#         toupper(strsplit(patches_string[x, ][[1]][[1]][patches_index],
-#                          " ")[[1]])
-#       patches_owns <- as.data.frame(matrix(patches_owns,
-#                                            nrow = 1,
-#                                            byrow = TRUE),
-#                                     stringsAsFactors = FALSE)
-#       patches_owns <- utils::type.convert(patches_owns)
-#
-#     }))
-#     names(patches_owns) <- getexp(nl, "metrics.patches")
-#
-#     return(patches_owns)
-#   })
-#   patches_owns
-# }
-#
-# .util_clean_metrics_links <- function(NLtable, nl) {
-#
-#   ## Extract current string:
-#   links_string <- NLtable[, grepl(c("metrics.links"), names(NLtable))]
-#
-#   links_owns <- purrr::map(seq_len(nrow(links_string)), function(x) {
-#
-#     ## Remove the outer [] brackets
-#     links_string[x, ][[1]] <- substring(links_string[x, ][[1]], 2)
-#     links_string[x, ][[1]] <- substring(links_string[x, ][[1]], 1,
-#                                         max(nchar(links_string[x, ][[1]])) - 1)
-#     links_string[x, ][[1]] <- gsub("breed", "", links_string[x, ][[1]],
-#                                    perl = TRUE)
-#
-#     links_string[x, ][[1]] <- gsub("^ *|(?<= ) | *$", "",
[…]
