+++ man/getexp.Rd
@@ -15,7 +15,16 @@ getexp(nl, var)
 Getter function to get a variable of an experiment object
 }
 \examples{
-\dontrun{
-getexp(nl, "expname")
-}
+# Example for Wolf Sheep Predation model from NetLogo models library:
+nl <- nl(nlversion = "6.0.3",
+nlpath = "/home/user/NetLogo 6.0.3/",
+modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
+jvmmem = 1024)
+
+# Set experiment name
+setexp(nl, "expname") <- "experimentName"
+
+# Get experiment name
+getexp(nl, "experiment")
+
 }
