### 2018-12-19 #262 @marcosci
Hi Anna, Marina and Thomas - thanks a ton for the very nice reviews!

Just trying to fix the issues @marinapapa had with the `get_nl_spatial` function - can you provide me with the script you had that caused the errors? :)
Just tried it with our toy examples (https://github.com/nldoc/nldoc_playground) and I couldn't reproduce that.
