### 2018-12-19 bed81f1
small cleaning of spatial output vignette
