+++ NEWS.md
@@ -1,0 +1,3 @@
+# nlrx 0.1.0
+
+* First release to CRAN.
