### 2018-12-19 55d098e
load libraries at same place in spatial vignette
