### 2018-12-20 5f8dc3d
Documentation updates Class constructor Readme Further notes
