### 2018-12-20 90c42ad
added more descriptive titles and details sections to util_eval functions
