+++ R/util_eval.R
@@ -1,7 +1,10 @@
-#' Evaluate variables list of an experiment object
+#' Evaluate if variables list of an experiment object is empty
 #'
-#' @description Evaluate variables list of an experiment object
+#' @description Evaluate if variables list of an experiment object is empty
 #' @param nl nl object
+#' @details
+#' util_eval_variables checks if the variables list of an experiment within a nl object is empty.
+#' It reports an error message if no variables were defined.
 #' @aliases util_eval_variables
 #' @rdname util_eval_variables
 #' @keywords internal
@@ -18,18 +21,17 @@ util_eval_variables <- function(nl) {
 
 
 
-#' Evaluate variables list of an experiment object
+#' Evaluate variables list of an experiment object for distinct simdesign
 #'
-#' @description Evaluate variables list of an experiment object
+#' @description Evaluate variables list of an experiment object for distinct simdesign
 #' @param nl nl object
+#' @details
+#' util_eval_variables_distinct checks if the variables list of an experiment within a nl object has enough information to create a \link[nlrx]{simdesign_distinct}.
+#' It reports an error message if at least one variable does not have a vector of distinct values or if there is a mismatch in length of these values vectors.
 #' @aliases util_eval_variables_distinct
 #' @rdname util_eval_variables_distinct
 #' @keywords internal
 util_eval_variables_distinct <- function(nl) {
-
-  # The function checks if there is enough variable information to create a ff
-  # simdesign
-  # The ff design needs either a set of min, max and step or distinct values
 
   vars <- getexp(nl, "variables")
   vars.values.missing <- lapply(vars, function(x) is.null(x$values))
@@ -72,10 +74,13 @@ util_eval_variables_distinct <- function(nl) {
 
 
 
-#' Evaluate variables list of an experiment object
+#' Evaluate variables list of an experiment object for full-factorial simdesign
 #'
-#' @description Evaluate variables list of an experiment object
+#' @description Evaluate variables list of an experiment object for full-factorial simdesign
 #' @param nl nl object
+#' @details
+#' util_eval_variables_ff checks if the variables list of an experiment within a nl object has enough information to create a \link[nlrx]{simdesign_ff}.
+#' It reports an error message if at least one variable does not have a defined sequence (min, max, step) or a vector of distinct values.
 #' @aliases util_eval_variables_ff
 #' @rdname util_eval_variables_ff
 #' @keywords internal
@@ -106,10 +111,13 @@ util_eval_variables_ff <- function(nl) {
 }
 
 
-#' Evaluate variables list of an experiment object
+#' Evaluate variables list of an experiment object for sensitivity analysis simdesigns
 #'
-#' @description Evaluate variables list of an experiment object
+#' @description Evaluate variables list of an experiment object for sensitivity analysis simdesigns
 #' @param nl nl object
+#' @details
+#' util_eval_variables_sa checks if the variables list of an experiment within a nl object has enough information to create a sensitivity analysis simdesign.
+#' It reports an error message if at least one variable does not have a defined distribution (min, max, qfun).
 #' @aliases util_eval_variables_sa
 #' @rdname util_eval_variables_sa
 #' @keywords internal
@@ -140,10 +148,13 @@ util_eval_variables_sa <- function(nl) {
 
 
 
-#' Evaluate variables list of an experiment object
+#' Evaluate variables list of an experiment object for optimization simdesigns
 #'
-#' @description Evaluate variables list of an experiment object
+#' @description Evaluate variables list of an experiment object for optimization simdesigns
 #' @param nl nl object
+#' @details
[…]
