### 2018-12-20 58bff80
util_eval_experiment: Added error message to warn if a netlogo variable is defined in variables and constants slot of an experiment
