+++ man/report_model_parameters.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/util_stuff.R
-\name{load_model_parameters}
-\alias{load_model_parameters}
+\name{report_model_parameters}
+\alias{report_model_parameters}
 \title{Report globals from a NetLogo model that is defined within a nl object}
 \usage{
-load_model_parameters(nl)
+report_model_parameters(nl)
 }
 \arguments{
 \item{nl}{nl object with a defined modelpath that points to a NetLogo model
@@ -23,7 +23,7 @@ entered into the variables or constants vector of an experiment object.
 }
 \examples{
 \dontrun{
-load_model_parameters(nl)
+report_model_parameters(nl)
 }
 
 }
