+++ NAMESPACE
@@ -15,8 +15,8 @@ export(getexp)
 export(getnl)
 export(getsim)
 export(import_nl)
-export(load_model_parameters)
 export(nl)
+export(report_model_parameters)
 export(run_nl_all)
 export(run_nl_dyn)
 export(run_nl_one)
