### 2018-12-20 73d914c
write_simoutput: 1. added warning if output tibble is empty with a hint that results need to be attached first 2. added error if output directory does not exist 3. added option to reassign the output directory
