### 2018-12-20 5d02790
analyze_nl, documentation: Added more information on analysis functions with links to the corresponding analysis sub-functions
