### 2018-12-20 1c837b6
Simdesign examples vignette: Added more descriptive explanations
