+++ R/class_constr.R
@@ -218,6 +218,8 @@ experiment <- function(expname = "defaultexp",
 #'
 #' \link[nlrx]{simdesign_lhs}
 #' The latin hypercube simdesign creates a Latin Hypercube sampling parameter matrix.
+#' The method can be used to generate a near-random sample of parameter values from the defined parameter distributions.
+#' More Details on Latin Hypercube Sampling can be found in McKay 1979 (https://doi.org/10.2307%2F1268522).
 #' To setup a latin hypercube sampling simdesign, variable distributions need to be defined (min, max, qfun).
 #'
 #' Sensitivity Analyses: \link[nlrx]{simdesign_sobol}, \link[nlrx]{simdesign_sobol2007}, \link[nlrx]{simdesign_soboljansen}, \link[nlrx]{simdesign_morris}, \link[nlrx]{simdesign_eFast}
