+++ vignettes/simdesign-examples.Rmd
@@ -51,7 +51,8 @@ nl <- nl(nlversion = "6.0.3",
 #### Simple simulation using constants (simdesign_simple)
 
 The simple simdesign only uses defined constants and reports a parameter matrix with only one parameterization.
-To setup a simple simdesign, no variables have to be defined. 
+To setup a simple simdesign, no variables have to be defined. The simple simdesign can be used to test a specific parameterisation of the model.
+It is also useful to for creating animated output of one specific parameterisation (see "Capturing Spatial NetLogo Output" vignette).
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
@@ -144,7 +145,10 @@ nl@simdesign <- simdesign_ff(nl=nl,
 
 #### Latin Hypercube Sampling (simdesign_lhs)
 
-
+The latin hypercube simdesign creates a Latin Hypercube sampling parameter matrix.
+The method can be used to generate a near-random sample of parameter values from the defined parameter distributions.
+More Details on Latin Hypercube Sampling can be found in [McKay 1979](https://doi.org/10.2307%2F1268522).
+To setup a latin hypercube sampling simdesign, variable distributions need to be defined (min, max, qfun).
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
@@ -181,7 +185,7 @@ nl@simdesign <- simdesign_lhs(nl=nl,
 #### Sensitivity Analyses (simdesign_sobol, _sobol2007, _soboljansen, _morris, _eFast)
 
 Sensitivity analyses are useful to estimate the importance of model parameters and to scan the parameter space in an efficient way.
-All supported sensitivity analysis simdesigns can be used to calculate sensitivity indices for each parameter-output combination. These indices can be calculated by using the analyze_nl() function after attaching the simulation results to the nl object.
+All supported sensitivity analysis simdesigns can be used to calculate sensitivity indices for each parameter-output combination. These indices can be calculated by using the analyze_nl() function after attaching the simulation results to the nl object. 
 To setup sensitivity analysis simdesigns, variable distributions (min, max, qfun) need to be defined.
 
 ```{r eval=FALSE}
