### 2018-12-20 530e8a4
Optimization Simdesigns and run_nl_dyn: Added notes, that optimization simdesign can only be executed with run_nl_dyn()
