+++ vignettes/simdesign-examples.Rmd
@@ -256,6 +256,7 @@ nl@simdesign <- simdesign_eFast(nl=nl,
 Optimization techniques are a powerful tool to search the parameter space for specific solutions.
 Both approaches try to minimize a specified model output reporter by systematically (genetic algorithm) or randomly (simulated annealing) changing the model parameters within the allowed ranges.
 To setup optimization simdesigns, variable ranges (min, max) need to be defined.
+Optimization simdesigns can only be executed using the `run_nl_dyn()` function instead of `run_nl_all()` or `run_nl_one()`.
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
