### 2018-12-20 0c11b4d
util_cleanup: Major change in how the cleanup function works. New version allows more control over which temporary files should not be deleted. All examples in all sections have been adapted accordingly.
