+++ R/nlrx-package.R
@@ -76,8 +76,7 @@
 #' \code{
 #' future::plan(multisession)
 #'
-#' results %<-% run_nl_all(nl = nl,
-#'                         cleanup = "all")
+#' results %<-% run_nl_all(nl = nl)
 #'
 #' }
 #'
