+++ R/run_nl.R
@@ -6,14 +6,19 @@
 #'
 #' @param nl nl object
 #' @param split number of parts the job should be split into
-#' @param cleanup indicate which filetypes should be deleted
+#' @param cleanup.csv TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.
+#' @param cleanup.xml TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.
+#' @param cleanup.bat TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.
 #' @return tibble with simulation output results
 #' @details
 #'
 #' run_nl_all executes all simulations of the specified NetLogo model within the provided nl object.
 #' The function loops over all random seeds and all rows of the siminput table of the simdesign of nl.
 #' The loops are created by calling \link[furrr]{future_map_dfr}, which allows running the function either locally or on remote HPC machines.
-#' Cleanup can either be ".xml" to delete all temporarily created xml files; ".csv" to delete all temporarily created csv files or "all" to delete all temporarily created files.
+#' The logical cleanup variables can be set to FALSE to preserve temporary generated output files (e.g. for debugging).
+#' cleanup.csv deletes/keeps the temporary generated model output files from each run.
+#' cleanup.xml deletes/keeps the temporary generated experiment xml files from each run.
+#' cleanup.bat deletes/keeps the temporary generated batch/sh commanline files from each run.
 #'
 #' When using run_nl_all in a parallelized environment (e.g. by setting up a future plan using the future package),
 #' the outer loop of this function (random seeds) creates jobs that are distributed to available cores of the current machine.
@@ -40,7 +45,11 @@
 #'
 #' @export
 
-run_nl_all <- function(nl, split = 1, cleanup = "all") {
+run_nl_all <- function(nl,
+                       split = 1,
+                       cleanup.csv = TRUE,
+                       cleanup.xml = TRUE,
+                       cleanup.bat = TRUE) {
   ## Store the number of siminputrows
   siminput_nrow <- nrow(getsim(nl, "siminput"))
   ## Check if split parameter is valid:
@@ -98,15 +107,19 @@ run_nl_all <- function(nl, split = 1, cleanup = "all") {
 #' @param nl nl object
 #' @param seed a random seed for the NetLogo simulation
 #' @param siminputrow rownumber of the input tibble within the attached simdesign object that should be executed
-#' @param cleanup indicate which filetypes should be deleted
+#' @param cleanup.csv TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.
+#' @param cleanup.xml TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.
+#' @param cleanup.bat TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.
 #' @return tibble with simulation output results
 #' @details
 #'
 #' run_nl_one executes one simulation of the specified NetLogo model within the provided nl object.
 #' The random seed is set within the NetLogo model to control stochasticity.
 #' The siminputrow number defines which row of the input data tibble within the simdesign object of the provided nl object is executed.
-#' Cleanup can either be "xml" to delete all temporarily created xml files; "csv" to delete all temporarily created csv files,
[…]
