### 2018-12-20 7e2aa14
util_cleanup: small bugfix
