### 2018-12-20 89eea8b
util_gather_results: 1. Added more informative error messages when runtime errors happen and no output has been created. 2. Added more informative error message when temporary csv output file was not found (this can happen on unix systems due to automatic deletion of temp files by the system)
