### 2018-12-20 b2b969e
further notes vignette renamed to advanced configuration; added section on temp file management
