+++ R/nlrx-package.R
@@ -6,77 +6,94 @@
 #'
 #' @details
 #'
-#' The nlrx package uses S4 classes to store basic information on NetLogo, the model experiment and the simulation design.
-#' Experiment and simulation design class objects are stored within the nl class object.
-#' This allows to have a complete simulation setup within one single R object.
+#' Get started:
 #'
-#' The following steps guide you trough the process on how to setup, run and analyze NetLogo model simulations with nlrx
+#' General information that is needed to run NetLogo simulations remotely, such as path to the NetLogo installation folder is stored within a `nl` class object.
+#' Nested within this `nl` class are the classes `experiment` and `simdesign`. The `experiment` class stores all experiment specifications.
+#' After attaching a valid experiment, a `simdesign` class object can be attached to the `nl` class object, by using one of the simdesign helper functions.
+#' These helper functions create different parameter input matrices from the experiment variable definitions that can then be executed by the `run_nl_one()` and `run_nl_all()` functions.
+#' The nested design allows to store everything related to the experiment within one R object.
+#' Additionally, different simdesign helper functions can be applied to the same `nl` object in order to repeat the same experiment with different parameter exploration methods (simdesigns).
+#'
+#' Step by step application example
+#'
+#' The "Wolf Sheep Predation" model from the NetLogo models library is used to present a basic example on how to setup and run NetLogo model simulations from R.
 #'
 #' Step 1: Create a nl object:
 #'
 #' The nl object holds all information on the NetLogo version, a path to the NetLogo directory with the defined version, a path to the model file, and the desired memory for the java virtual machine.
+#' Depending on the operation system, paths to NetLogo and the model need to be adjusted.
 #'
 #' \code{
+#' library(nlrx)
+#' # Windows default NetLogo installation path (adjust to your needs!):
+#' netlogopath <- file.path("C:/Program Files/NetLogo 6.0.3")
+#' modelpath <- file.path(netlogopath, "app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
+#' outpath <- file.path("C:/out")
+#' # Unix default NetLogo installation path (adjust to your needs!):
+#' netlogopath <- file.path("/home/NetLogo 6.0.3")
+#' modelpath <- file.path(netlogopath, "app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
+#' outpath <- file.path("/home/out")
+#'
 #' nl <- nl(nlversion = "6.0.3",
-#' nlpath = "C:/Program Files/NetLogo 6.0.3/",
-#' modelpath = "C:/Program Files/NetLogo 6.0.3/app/models/
-#' Sample Models/Biology/Wolf Sheep Predation.nlogo",
-#' jvmmem = 1024)
+#'          nlpath = netlogopath,
+#'          modelpath = modelpath,
+#'          jvmmem = 1024)
 #' }
 #'
 #' Step 2: Attach an experiment
 #'
-#' The experiment object is organized in a similar fashion as NetLogo BehaviorSpace experiments.
-#' It holds information on model variables, constants, metrics, runtime, ...
+#' The experiment object is organized in a similar fashion as NetLogo Behavior Space experiments.
+#' It contains all information that is needed to generate a simulation parameter matrix and to execute the NetLogo simulations.
+#' Details on the specific slots of the experiment class can be found in the package documentation (`?experiment`) and the "Advanced configuration" vignette.
 #'
 #' \code{
 #' nl@@experiment <- experiment(expname="wolf-sheep",
-#' outpath="C:/out/",
-#' repetition=1,
-#' tickmetrics="true",
-#' idsetup="setup",
-#' idgo="go",
[…]
