+++ vignettes/articles/interfacing-different-models.Rmd
@@ -13,21 +13,364 @@ many different applications.
 All the following examples show how to create a gif animation from a NetLogo model.
 This always higlights how to setup very different model types.
 
-### [nlrx_ants_gif.R](https://github.com/nldoc/nlrx_playground/blob/master/nlrx_ants_gif.R)
+First, we need to load a bunch of packages, mainly for the subsetting of the spatial data
+and the plotting:
 
-![Alt Text](https://github.com/nldoc/nlrx_playground/blob/master/nlrx_ants_gif.gif)
+```{r eval = FALSE}
+## Load packages
+library(nlrx)
+library(tidyverse)
+library(ggplot2)
+library(gganimate)
+library(emojifont) 
+library(cartography) 
+```
+
+... and now are ready to interface some of the models of the NetLogo model library.  If you have problems to interface a specific type of model, feel free to send one of the authors of nlrx an email!
+
+### Ants model
+```{r eval = FALSE}
+## Step1: Create a nl obejct:
+nl <- nl(nlversion = "6.0.4",
+         nlpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/",
+         modelpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/app/models/Sample Models/Biology/Ants.nlogo",
+         jvmmem = 1024)
+
+## Step2: Add Experiment
+nl@experiment <- experiment(expname = "nlrxtest",
+                            outpath = "C:/out",
+                            repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
+                            tickmetrics = "true",
+                            idsetup = "setup",   # you can define multiple setup procedures with c()
+                            idgo = "go",         # you can define multiple go procedures with c()
+                            runtime = 1000,
+                            evalticks = seq(1,1000),
+                            metrics.turtles = c("who", "pxcor", "pycor", "breed"),
+                            metrics.patches = c("pxcor", "pycor", "pcolor", "chemical", "food", "food-source-number"),
+                            constants = list("population" = 125,
+                                             'diffusion-rate' = 50,
+                                             'evaporation-rate' = 10))
+
+## Step3: Add a Simulation Design
+nl@simdesign <- simdesign_simple(nl = nl,
+                                 nseeds = 1)
 
 
-### [nlrx_diffusion_gif.R](https://github.com/nldoc/nlrx_playground/blob/master/nlrx_diffusion_gif.R)
+## Step4: Run simulations:
+results <- run_nl_all(nl = nl)
 
-![Alt Text](https://github.com/nldoc/nlrx_playground/blob/master/diffusion.gif)
+## Step5: Attach results to nl and reformat spatial data with get_nl_spatial()
+setsim(nl, "simoutput") <- results
+nl_spatial <- get_nl_spatial(nl, format = "tibble")
+
+## Step6: Prepare data for plotting
+nmax <- 900
+food_dat <- nl_spatial %>% dplyr::filter(food > 0) %>% dplyr::filter(step %in% 1:nmax) %>% dplyr::select(patches_x, patches_y, step)
+chem_dat <- nl_spatial %>% dplyr::filter(step %in% 1:nmax) %>% dplyr::select(patches_x, patches_y, chemical, step)
+chem_dat$chemical <- ifelse(chem_dat$chemical <= 0.2, NA, chem_dat$chemical)
+turt_dat <- nl_spatial %>% dplyr::filter(!is.na(who)) %>% dplyr::filter(step %in% 1:nmax) %>% dplyr::select(pxcor, pycor, who, step)
+nest_dat <- data.frame(x=rep(0, nmax), y=rep(0, nmax), step=1:nmax)
+
[…]
