+++ CONTRIBUTING.md
@@ -15,7 +15,7 @@ We appreciate any contribution and collaboration.
 
 * Fork this repo to your Github account
 * Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/nldoc/nlrx.git`
-* Make sure to track progress upstream (i.e., on our version of `landscapemetrics` at `r-spatialecology/landscapemetrics`) by doing `git remote add upstream https://github.com/nldoc/nlrx.git`.
+* Make sure to track progress upstream (i.e., on our version of `nlrx` at `nldoc/nlrx`) by doing `git remote add upstream https://github.com/nldoc/nlrx.git`.
 Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
 * Make your changes (bonus points for making changes on a new branch)
 * If you alter package functionality at all (e.g., the code itself, not just documentation) please do write some tests to cover the new functionality
