+++ _pkgdown.yml
@@ -5,7 +5,7 @@ template:
    bootswatch: sandstone
 
 navbar:
-  title: landscapemetrics
+  title: nlrx
   left:
     - text: "Get Started"
       icon: fa fa-magic
