+++ vignettes/articles/interfacing-different-models.Rmd
@@ -26,19 +26,31 @@ library(emojifont)
 library(cartography) 
 ```
 
+Second, we need to define the path variables pointing to the NetLogo installation.
+Depending on the operation system, the NetLogo and output paths need to be adjusted.
+
+```{r eval = FALSE}
+# Windows default NetLogo installation path (adjust to your needs!):
+netlogopath <- file.path("C:/Program Files/NetLogo 6.0.4")
+outpath <- file.path("C:/out")
+# Unix default NetLogo installation path (adjust to your needs!):
+netlogopath <- file.path("/home/NetLogo 6.0.4")
+outpath <- file.path("/home/out")
+```
+
 ... and now are ready to interface some of the models of the NetLogo model library.  If you have problems to interface a specific type of model, feel free to send one of the authors of nlrx an email!
 
 ## Ants model
 ```{r eval = FALSE}
 ## Step1: Create a nl obejct:
 nl <- nl(nlversion = "6.0.4",
-         nlpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/",
-         modelpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/app/models/Sample Models/Biology/Ants.nlogo",
+         nlpath = netlogopath,
+         modelpath = file.path(netlogopath, "app/models/Sample Models/Biology/Ants.nlogo"),
          jvmmem = 1024)
 
 ## Step2: Add Experiment
-nl@experiment <- experiment(expname = "nlrxtest",
-                            outpath = "C:/out",
+nl@experiment <- experiment(expname = "ants",
+                            outpath = outpath,
                             repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
                             tickmetrics = "true",
                             idsetup = "setup",   # you can define multiple setup procedures with c()
@@ -97,13 +109,13 @@ gganimate::animate(p1, nframes = max(food_dat$step), width=800, height=800, fps=
 ```{r  eval = FALSE}
 ## Step1: Create a nl obejct:
 nl <- nl(nlversion = "6.0.4",
-         nlpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/",
-         modelpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/app/models/Sample Models/Art/Diffusion Graphics.nlogo",
+         nlpath = netlogopath,
+         modelpath = file.path(netlogopath, "app/models/Sample Models/Art/Diffusion Graphics.nlogo"),
          jvmmem = 1024)
 
 ## Step2: Add Experiment
 nl@experiment <- experiment(expname = "diffusion",
-                            outpath = "/home/marco/Desktop/",
+                            outpath = outpath,
                             repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
                             tickmetrics = "true",
                             idsetup = "setup",   # you can define multiple setup procedures with c()
@@ -167,13 +179,13 @@ gganimate::animate(p1, width=800, height=800, duration=6)
 ```{r  eval = FALSE}
 ## Step1: Create a nl obejct:
 nl <- nl(nlversion = "6.0.4",
-         nlpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/",
-         modelpath = "/home/marco/Documents/rpackages/nlrx_usecase/1_Helper/NetLogo 6.0.4/app/models/Sample Models/Art/Diffusion Graphics.nlogo",
+         nlpath = netlogopath,
+         modelpath = file.path(netlogopath, "app/models/Sample Models/Art/Diffusion Graphics.nlogo"),
          jvmmem = 1024)
 
[…]
