### 2018-12-27 04f8715
fix test_util_runnl
