+++ _pkgdown.yml
@@ -34,4 +34,4 @@ navbar:
       href: reference/index.html
     - icon: fa-github fa-lg
       text: "github"
-      href: https://github.com/r-spatialecology/landscapemetric
+      href: https://github.com/nldoc/nlrx
