### 2019-01-05 #262 @annakrystalli
Hello all and happy (belated) New Year! 💫

Just checking in to see how things are going.

I also asked the editors about your query regarding the repeating warning message when running `get_nl_spatial` in Rmd. While not necessarily more urgent than other issues,it still needs to be addressed prior to approval.

@noamross also provided some useful tips on how to troubleshoot and then set up a test to monitor the behaviour:

> Debugging this thing can be tricky, but a good place to start is to make the minimal Rmd that produces and render it using a _script_ with `rmarkdown::render()`.  Then `debug()`, `options(error=recover)`, etc are all available to you., and then you can write a test to check that it renders properly in the future.

Let me know if you have any further questions!
