### 2019-01-05 #262 @nldoc
Happy New Year to you too!

Let me give you a short update on the progress of our revisions.
We revised large parts of the package documentation and already resolved most of the issues mentioned by the reviewers. There are still some small things to be done (for example the get_nl_spatial warning messages; thank you @annakrystalli for your hints, we will try to solve this issue asap).
We will continue fixing the remaining issues on Monday and I think we are ready to post our detailed response to all reviewer comments by midweek.
Have a nice weekend!
