### 2019-01-07 de13179
documentation fixes
