+++ man/analyze_eFast.Rd
@@ -16,4 +16,7 @@ analyze_eFast(nl, metrics, funs)
 \description{
 Analyze NetLogo simulation output of simdesign eFast
 }
+\details{
+The function calculates eFast sensitivity indices from the output results using the \link[sensitivity]{sensitivity} package.
+}
 \keyword{internal}
