### 2019-01-07 2670a16
Fixing error and warning messages in util_eval
