### 2019-01-07 671913e
Added notes on eval_variables_constants() and report_model_parameters() to Advanced Configuration Vignette
