### 2019-01-07 b995b6c
set .initClasses as internal
