+++ R/class_init.R
@@ -1,15 +1,12 @@
+
+
+#' @title initClasses
+#' @description empty
+#' @name init-classes
+#' @rdname init-classes
+#' @keywords internal
 .initClasses <- function() {
 
-  #' @title S4 class \code{"simdesign"}
-  #' @description empty
-  #' @slot simmethod A character string for method identification
-  #' @slot siminput Input parameter tibble for model simulations
-  #' @slot simobject Store method object (e.g. sobol, morris, eFast)
-  #' @slot simseeds A vector of model seeds
-  #' @slot simoutput variable to attach simulation result tibbles
-  #' @name simdesign-class
-  #' @rdname simdesign-class
-  #' @export
   simdesign <- methods::setClass("simdesign",
     slots = list(
       simmethod = "character",
@@ -28,51 +25,6 @@
     )
   )
 
-  #' @title S4 class \code{"experiment"}
-  #' @description empty
-  #' @slot expname A character string defining the name of the experiment
-  #' @slot outpath Path to a directory where experiment output will be stored
-  #' @slot repetition The number of repetitions for each row of the simulation
-  #' design input tibble
-  #' @slot tickmetrics "true" runs defined metrics on each simulation tick.
-  #' "false" runs metrics only after simulation is finished
-  #' @slot idsetup character string or vector of character strings, defining
-  #' the name of the NetLogo setup procedure
-  #' @slot idgo character string or vector of character strings, defining the
-  #'  name of the NetLogo go procedure
-  #' @slot idfinal character string or vector of character strings, defining
-  #' the name of NetLogo procedures that should be run after the last tick
-  #' @slot idrunnum character string, defining the name of a NetLogo global
-  #' that should be used to parse the current siminputrow during model
-  #' executions which can then be used for self-written output.
-  #' @slot runtime number of model ticks that should be run for each simulation
-  #' @slot evalticks number or vector of tick numbers defining when measurements
-  #'  are taken
-  #' @slot stopcond a NetLogo reporter that reports TRUE/FALSE. If it reports
-  #' TRUE the current simulation run is stopped (e.g. "not any? turtles")
-  #' @slot metrics vector of strings defining valid NetLogo reporters that are
-  #' taken as output measurements (e.g. c("count turtles", "count patches"))
-  #' @slot metrics.turtles vector of strings defining valid turtles-own
-  #' variables that are taken as output measurements (e.g. c("who", "pxcor",
-  #'  "pycor", "color"))
-  #' @slot metrics.patches vector of strings defining valid patches-own
-  #' variables that are taken as output measurements (e.g. c("pxcor",
-  #' "pycor", "pcolor"))
-  #' @slot metrics.links vector of strings defining valid links-own variables
-  #' that are taken as output measurements (e.g. c("end1", "end2"))
-  #' @slot variables a nested list of variables that are changed within a
-  #' simulation design. The name of each sublist item has to be a valid global
-  #' of the defined NetLogo model. Each list item consist of a min value, a max
-  #' value, a step value and a qfun (e.g. list("paramA" = list(min=0, max=1,
-  #' step=0.1, qfun="qunif")))
-  #' @slot constants a list of constants that are kept constant within a
-  #' simulation design. The name of each list item has to be a valid global of
-  #' the defined NetLogo model (e.g. list("pNUM" = 12, "pLOGIC"="TRUE",
-  #' "pSTRING"="\"default\""))
-  #' @slot simdesign Holds a simdesign S4 class object
-  #' @name experiment-class
-  #' @rdname experiment-class
-  #' @export
   experiment <- methods::setClass("experiment",
     slots = list(
       expname = "character",
[…]
