+++ man/init-classes.Rd
@@ -1,0 +1,13 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/class_init.R
+\name{init-classes}
+\alias{init-classes}
+\alias{.initClasses}
+\title{initClasses}
+\usage{
+.initClasses()
+}
+\description{
+empty
+}
+\keyword{internal}
