+++ NAMESPACE
@@ -4,7 +4,6 @@ export("%>%")
 export("setexp<-")
 export("setnl<-")
 export("setsim<-")
-export(.initClasses)
 export(analyze_nl)
 export(download_netlogo)
 export(eval_variables_constants)
