### 2019-01-07 92c7776
Fix tests util_eval
