### 2019-01-07 836efb7
Fix line length of experiment.Rd
