+++ man/experiment.Rd
@@ -150,13 +150,21 @@ nl@experiment <- experiment(expname="wolf-sheep",
                              runtime=50,
                              evalticks=seq(40,50),
                              stopcond="not any? turtles",
-                             metrics=c("count sheep", "count wolves", "count patches with [pcolor = green]"),
-                             metrics.turtles=c("who", "pxcor", "pycor", "color"),
+                             metrics=c("count sheep",
+                                       "count wolves",
+                                     "count patches with [pcolor = green]"),
+                             metrics.turtles=c("who",
+                                               "pxcor",
+                                               "pycor",
+                                               "color"),
                              metrics.patches=c("pxcor", "pycor", "pcolor"),
                              metrics.links=c("end1","end2"),
-                             variables = list('initial-number-sheep' = list(min=50, max=150, step=10, qfun="qunif"),
-                 '                             initial-number-wolves' = list(min=50, max=150, step=10, qfun="qunif")),
-                             constants = list("model-version" = "\\"sheep-wolves-grass\\"",
+                             variables = list('initial-number-sheep' =
+                             list(min=50, max=150, step=10, qfun="qunif"),
+                                              'initial-number-wolves' =
+                             list(min=50, max=150, step=10, qfun="qunif")),
+                             constants = list("model-version" =
+                                              "\\"sheep-wolves-grass\\"",
                                               "grass-regrowth-time" = 30,
                                               "sheep-gain-from-food" = 4,
                                               "wolf-gain-from-food" = 20,
