### 2019-01-07 5d86845
Added progress option to run_nl_X functions and updated documentation accordingly
