+++ man/run_nl_one.Rd
@@ -5,7 +5,7 @@
 \title{Execute one NetLogo simulation from a nl object}
 \usage{
 run_nl_one(nl, seed, siminputrow, cleanup.csv = TRUE,
-  cleanup.xml = TRUE, cleanup.bat = TRUE)
+  cleanup.xml = TRUE, cleanup.bat = TRUE, silent = TRUE)
 }
 \arguments{
 \item{nl}{nl object}
@@ -19,6 +19,8 @@ run_nl_one(nl, seed, siminputrow, cleanup.csv = TRUE,
 \item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}
 
 \item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}
+
+\item{silent}{TRUE/FALSE, if FALSE prints the current seed and siminputrow after successful execution of a simulation (only for sequential execution).}
 }
 \value{
 tibble with simulation output results
