+++ man/util_run_nl_dyn_GenAlg.Rd
@@ -4,7 +4,8 @@
 \alias{util_run_nl_dyn_GenAlg}
 \title{Genetic Algorithm call simulations function}
 \usage{
-util_run_nl_dyn_GenAlg(nl, seed, cleanup.csv, cleanup.xml, cleanup.bat)
+util_run_nl_dyn_GenAlg(nl, seed, cleanup.csv, cleanup.xml, cleanup.bat,
+  silent)
 }
 \arguments{
 \item{nl}{nl object}
@@ -16,6 +17,8 @@ util_run_nl_dyn_GenAlg(nl, seed, cleanup.csv, cleanup.xml, cleanup.bat)
 \item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}
 
 \item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}
+
+\item{silent}{TRUE/FALSE, if FALSE prints the current seed and siminputrow after successful execution of a simulation (only for sequential execution).}
 }
 \description{
 Genetic Algorithm call simulations function

[…]
