### 2019-01-07 cd6814d
fix some spellings
