+++ R/run_nl.R
@@ -19,7 +19,7 @@
 #' The logical cleanup variables can be set to FALSE to preserve temporary generated output files (e.g. for debugging).
 #' cleanup.csv deletes/keeps the temporary generated model output files from each run.
 #' cleanup.xml deletes/keeps the temporary generated experiment xml files from each run.
-#' cleanup.bat deletes/keeps the temporary generated batch/sh commanline files from each run.
+#' cleanup.bat deletes/keeps the temporary generated batch/sh commandline files from each run.
 #'
 #' When using run_nl_all in a parallelized environment (e.g. by setting up a future plan using the future package),
 #' the outer loop of this function (random seeds) creates jobs that are distributed to available cores of the current machine.
@@ -125,7 +125,7 @@ run_nl_all <- function(nl,
 #' The logical cleanup variables can be set to FALSE to preserve temporary generated output files (e.g. for debugging).
 #' cleanup.csv deletes/keeps the temporary generated model output files from each run.
 #' cleanup.xml deletes/keeps the temporary generated experiment xml files from each run.
-#' cleanup.bat deletes/keeps the temporary generated batch/sh commanline files from each run.
+#' cleanup.bat deletes/keeps the temporary generated batch/sh commandline files from each run.
 #'
 #' This function can be used to run single simulations of a NetLogo model.
 #'
@@ -211,7 +211,7 @@ run_nl_one <- function(nl,
 #' The logical cleanup variables can be set to FALSE to preserve temporary generated output files (e.g. for debugging).
 #' cleanup.csv deletes/keeps the temporary generated model output files from each run.
 #' cleanup.xml deletes/keeps the temporary generated experiment xml files from each run.
-#' cleanup.bat deletes/keeps the temporary generated batch/sh commanline files from each run.
+#' cleanup.bat deletes/keeps the temporary generated batch/sh commandline files from each run.
 #'
 #' @examples
 #' \dontrun{
