+++ vignettes/simdesign-examples.Rmd
@@ -83,7 +83,7 @@ nl@simdesign <- simdesign_simple(nl=nl,
 
 #### Distinct parameter combinations (simdesign_distinct)
 
-The distinct simdesign can be used to run distinct parameter combinations. To setup a distinct simdesign, vectors of values need to be defined for each variable. These vectors must have the same number of elemtents across all variables. The first simulation run consist of all 1st elements of these variable vectors; the second run uses all 2nd values, and so on.
+The distinct simdesign can be used to run distinct parameter combinations. To setup a distinct simdesign, vectors of values need to be defined for each variable. These vectors must have the same number of elements across all variables. The first simulation run consist of all 1st elements of these variable vectors; the second run uses all 2nd values, and so on.
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
