+++ vignettes/furthernotes.Rmd
@@ -58,7 +58,7 @@ However, both approaches only work for sequential execution. Capturing output fr
 
 #### Handling NetLogo runtime errors
 
-Usually, runtime errors of NetLogo model simulations are printed to the R console and the current execution of run_nl_all(), run_nl_one() or run_nl_dyn() breaks. However, it can happen that a model simulation freezes due to Java runtime errors. Unfortunately it is not possible to terminate the Java virtual machine or print an error message to the console after such a runtime error occured.
+Usually, runtime errors of NetLogo model simulations are printed to the R console and the current execution of run_nl_all(), run_nl_one() or run_nl_dyn() breaks. However, it can happen that a model simulation freezes due to Java runtime errors. Unfortunately it is not possible to terminate the Java virtual machine or print an error message to the console after such a runtime error occurred.
 The current R session and the freezed Java Virtual Machine need to be terminated manually.
 Thus, NetLogo models should be debugged in NetLogo prior to execution of large experiments with nlrx.
 Capturing progress of model simulations (see section "Capturing progress of model simulations") might help in debugging runtime errors that freeze the Java Virtual Machine.
