### 2019-01-08 3a290e0
tidy up some bits and pieces
