+++ man/run_nl_dyn.Rd
@@ -32,7 +32,7 @@ This is the case for dynamic designs, such as Simulated Annealing and Genetic Al
 The logical cleanup variables can be set to FALSE to preserve temporary generated output files (e.g. for debugging).
 cleanup.csv deletes/keeps the temporary generated model output files from each run.
 cleanup.xml deletes/keeps the temporary generated experiment xml files from each run.
-cleanup.bat deletes/keeps the temporary generated batch/sh commanline files from each run.
+cleanup.bat deletes/keeps the temporary generated batch/sh commandline files from each run.
 }
 \examples{
 \dontrun{
