+++ vignettes/simdesign-examples.Rmd
@@ -19,8 +19,8 @@ knitr::opts_chunk$set(
 
 ## Simdesign examples
 
-nlrx provides a bunch of different simdesigns that generate parameter input matrices from valid experiment variable definitions.
-These simdesigns represent different techniques of parameter space exploration.
+nlrx provides a variety of different simdesigns that generate parameter input matrices from valid experiment variable definitions.
+These simdesigns represent different techniques of parameter space explorations.
 This is not a technical guide on parameter space exploration but rather a collection of code examples on how to setup these different approaches with nlrx.
 Details on the sensitivity analysis methods can be found in the documentation of the [sensitivity package](https://cran.r-project.org/web/packages/sensitivity/index.html).
 Details on the optimization methods can be found in the documentation of the [genalg package](https://cran.r-project.org/web/packages/genalg/index.html) and the [GenSA package](https://cran.r-project.org/web/packages/GenSA/index.html).
@@ -53,6 +53,9 @@ nl <- nl(nlversion = "6.0.3",
 The simple simdesign only uses defined constants and reports a parameter matrix with only one parameterization.
 To setup a simple simdesign, no variables have to be defined. The simple simdesign can be used to test a specific parameterisation of the model.
 It is also useful to for creating animated output of one specific parameterisation (see "Capturing Spatial NetLogo Output" vignette).
+
+`simdesign_simple` is most useful, if you want to imitate pressing the _go_ button in your model in NetLogo - just a single run, with a
+defined parameterisation.
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
@@ -110,7 +113,6 @@ nl@experiment <- experiment(expname="wolf-sheep",
 nl@simdesign <- simdesign_distinct(nl=nl,
                                    nseeds=3)
 ```
-
 
 #### Full-factorial simulation (simdesign_ff)
 
@@ -187,7 +189,7 @@ nl@simdesign <- simdesign_lhs(nl=nl,
 
 Sensitivity analyses are useful to estimate the importance of model parameters and to scan the parameter space in an efficient way.
 nlrx uses the [sensitivity](https://cran.r-project.org/web/packages/sensitivity/index.html) package to setup sensitivity analysis parameter matrices.
-All supported sensitivity analysis simdesigns can be used to calculate sensitivity indices for each parameter-output combination. These indices can be calculated by using the analyze_nl() function after attaching the simulation results to the nl object. 
+All supported sensitivity analysis simdesigns can be used to calculate sensitivity indices for each parameter-output combination. These indices can be calculated by using the `analyze_nl()` function after attaching the simulation results to the nl object. 
 To setup sensitivity analysis simdesigns, variable distributions (min, max, qfun) need to be defined.
 
 ```{r eval=FALSE}
@@ -258,7 +260,7 @@ nl@simdesign <- simdesign_eFast(nl=nl,
 Optimization techniques are a powerful tool to search the parameter space for specific solutions.
 Both approaches try to minimize a specified model output reporter by systematically (genetic algorithm, utilizing the [genalg](https://cran.r-project.org/web/packages/genalg/index.html) package) or randomly (simulated annealing, utilizing the [genSA](https://cran.r-project.org/web/packages/GenSA/index.html) package) changing the model parameters within the allowed ranges.
 To setup optimization simdesigns, variable ranges (min, max) need to be defined.
-Optimization simdesigns can only be executed using the `run_nl_dyn()` function instead of `run_nl_all()` or `run_nl_one()`.
+Optimization simdesigns can only be executed using the `run_nl_dyn()` function - instead of `run_nl_all()` or `run_nl_one()`.
 
 ```{r eval=FALSE}
 nl@experiment <- experiment(expname="wolf-sheep",
