### 2019-01-08 a1a21eb
Homepage update, typos etc.
