+++ vignettes/articles/spatial-output.Rmd
@@ -79,12 +79,12 @@ nl@simdesign <- simdesign_simple(nl=nl,
 results <- run_nl_all(nl = nl)
 ```
 
-This experiment will run for 100 ticks and collects all metrics, metrics.turtles and metrics.patches on each tick.
-Thus, running run_nl_all() will report a tibble containing all metrics, metrics.turtles and metrics.patches. However, because the spatial metrics contain more than one value, these datasets are stored as lists inside the output tibble.
-This lists already contain all measured agent metrics and can for example be used to analyse distributions of these variables for specific agent groups.
+This experiment will run for 100 ticks (runtime) and collects all metrics, metrics.turtles and metrics.patches on each tick (evalticks).
+Thus, executing `run_nl_all()` will report a tibble containing all metrics, metrics.turtles and metrics.patches. However, because the spatial metrics contain more than one value, these datasets are stored as lists inside the output tibble.
+These lists already contain all measured agent metrics and can for example be used to analyze distributions of these variables for specific agent groups.
 
 In case of spatial data containing coordinates (pxcor/pycor for patches or pxcor/pycor and/or xcor/ycor for turtles), nlrx provides a function to transform the measured output into spatial objects.
-In order to use the function get_nl_spatial() you first have to attach your simulation output results to the nl object:
+In order to use the function `get_nl_spatial()` the simulation output results tibble needs to be attached to the nl object first. The simdesign class within the nl object provides a slot for attaching output results (simoutput). An output results tibble can be attached to this slot by using the simdesign setter function `setsim(nl, "simoutput")`.
 
 ```{r eval=FALSE}
 # Attach results to nl object:
@@ -98,16 +98,16 @@ results_spatial <- get_nl_spatial(nl,
                                   format="spatial")
 ``` 
 
-The get_nl_spatial() function uses 4 function parameters:
+The `get_nl_spatial()` function uses 4 function parameters:
 
 * turtles - TRUE/FALSE, if true the function will transform the measured metrics.turtles to spatial sf objects.
 * patches - TRUE/FALSE, if true the function will transform the measured metrics.patches to spatial raster objects. In case there are several patches variables besides the coordinates, the result will be a rasterstack containing one raster for each patch variable
 * turtle_coords - "px"/"x", because turtle coordinates can be measured as pxcor/pycor or xcor/ycor this string defines which coordinates are used for creating the turtle sf objects. For instance, you can measure both coordinate pairs by entering them into metrics.turtles, and then decide for one type of coordinates to create spatial objects.
-* format - "spatial"/"tibble", the "spatial" option will report the generated output as tibble containing the spatial objects (raster/rasterStack, sf). You can also choose format "tibble" report the spatial objects as a long format tibble, that can be used for plotting landscapes and turtles with ggplot2. 
+* format - "spatial"/"tibble", the "spatial" option will report the generated output as tibble containing the spatial objects (raster/rasterStack, sf). You can also choose format "tibble" to report the spatial objects as a long format tibble, that can be used for plotting landscapes and turtles with ggplot2. 
 
 ## Visualization examples
 
-The spatial tibble output from get_nl_spatial() can for example be used to plot maps for different ticks of the model simulation.
+The spatial tibble output from `get_nl_spatial()` can for example be used to plot maps for different ticks of the model simulation.
 Here is an example to create a facet plot using spatial simulation data of every 10th simulation tick:
 
 ```{r eval=FALSE}
@@ -143,7 +143,7 @@ ggplot() +
 <img src="wolfsheep_world.png" align="center" width="100%" />
 
 Using the gganimate package(https://github.com/thomasp85/gganimate), it is even possible to generate animated plots from this spatial data tibble.
[…]

[…]
