### 2019-01-09 7a9104c
Interfacing vignette: added links example
