+++ man/run_nl_one.Rd
@@ -44,7 +44,7 @@ This function can be used to run single simulations of a NetLogo model.
 
 # Run one simulation:
 results <- run_nl_one(nl = nl,
-                      seed = getsims(nl, "simseeds")[1],
+                      seed = getsim(nl, "simseeds")[1],
                       siminputrow = 1)
 
 }
