+++ vignettes/articles/interfacing-different-models.Rmd
@@ -59,12 +59,10 @@ nl <- nl(nlversion = "6.0.4",
 ## Step2: Add Experiment
 nl@experiment <- experiment(expname = "diffusion",
                             outpath = outpath,
-                            repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
+                            repetition = 1,   
                             tickmetrics = "true",
-                            idsetup = "setup",   # you can define multiple setup procedures with c()
-                            idgo = "go",         # you can define multiple go procedures with c()
-                            idfinal = NA_character_,  # you can define one or more final commands here
-                            idrunnum = NA_character_,
+                            idsetup = "setup",  
+                            idgo = "go",        
                             runtime = 500,
                             evalticks = seq(1,500),
                             constants = list("num-turtles" = 20,
@@ -135,12 +133,10 @@ nl <- nl(nlversion = "6.0.4",
 ## Step2: Add Experiment
 nl@experiment <- experiment(expname = "diffusion",
                             outpath = outpath,
-                            repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
+                            repetition = 1,    
                             tickmetrics = "true",
-                            idsetup = "setup",   # you can define multiple setup procedures with c()
-                            idgo = "go",         # you can define multiple go procedures with c()
-                            idfinal = NA_character_,  # you can define one or more final commands here
-                            idrunnum = NA_character_,
+                            idsetup = "setup", 
+                            idgo = "go",       
                             runtime = 500,
                             evalticks = seq(1,500),
                             metrics = c("count patches"),
@@ -230,10 +226,10 @@ nl <- nl(nlversion = "6.0.4",
 ## Step2: Add Experiment
 nl@experiment <- experiment(expname = "ants",
                             outpath = outpath,
-                            repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
+                            repetition = 1,      
                             tickmetrics = "true",
-                            idsetup = "setup",   # you can define multiple setup procedures with c()
-                            idgo = "go",         # you can define multiple go procedures with c()
+                            idsetup = "setup",   
+                            idgo = "go",         
                             runtime = 1000,
                             evalticks = seq(1,1000),
                             metrics.turtles = c("who", "pxcor", "pycor", "breed"),
@@ -304,18 +300,15 @@ nl <- nl(nlversion = "6.0.4",
 ## Step2: Add Experiment
 nl@experiment <- experiment(expname = "waves",
                             outpath = outpath,
-                            repetition = 1,      # If repetitions is > 1, a different random seed will be set for each netlogo run
+                            repetition = 1,      
                             tickmetrics = "true",
-                            idsetup = "setup",   # you can define multiple setup procedures with c()
-                            idgo = "go",         # you can define multiple go procedures with c()
-                            idfinal = NA_character_,  # you can define one or more final commands here
-                            idrunnum = NA_character_,
+                            idsetup = "setup",  
+                            idgo = "go",
                             runtime = 100,
                             evalticks = seq(1,100),
                             variables = list('friction' = list(values = c(5,25,50,90))),
[…]
