### 2019-01-09 e5f52d1
update get_nl_spatial
