### 2019-01-09 92497ca
fix get nl spatial
