### 2019-01-09 6d41329
fix fix get nl spatial
