### 2019-01-09 009480c
suppress messages in joins for get_nl_spatial
