+++ R/get_nl_spatial.R
@@ -150,21 +150,21 @@ get_nl_spatial <- function(nl,
     {
       turtles_tib$group <- "turtles"
       patches_tib$group <- "patches"
-      agentdata <- dplyr::full_join(patches_tib, turtles_tib)
+      agentdata <- suppressMessages(dplyr::full_join(patches_tib, turtles_tib))
     } else if (all(isTRUE(turtles) && isTRUE(links) && !isTRUE(patches) )) {
       turtles_tib$group <- "turtles"
       links_tib$group <- "links"
-      agentdata <- dplyr::full_join(links_tib, turtles_tib)
+      agentdata <- suppressMessages(dplyr::full_join(links_tib, turtles_tib))
     } else if (all(isTRUE(links) && isTRUE(patches ) && !isTRUE(turtles) )) {
       links_tib$group <- "links"
       patches_tib$group <- "patches"
-      agentdata <- dplyr::full_join(patches_tib, turtles_tib)
+      agentdata <- suppressMessages(dplyr::full_join(patches_tib, turtles_tib))
     } else if (all(isTRUE(turtles) && isTRUE(patches) && isTRUE(links))) {
       turtles_tib$group <- "turtles"
       patches_tib$group <- "patches"
       links_tib$group <- "links"
-      agentdata <- dplyr::full_join(patches_tib, turtles_tib) %>%
-        dplyr::full_join(links_tib)
+      agentdata <- suppressMessages(dplyr::full_join(patches_tib, turtles_tib) %>%
+        dplyr::full_join(links_tib))
     }
   }
 
@@ -258,7 +258,7 @@ get_nl_spatial <- function(nl,
 
     if (all(isTRUE(turtles) && isTRUE(patches)))
     {
-      agentdata <-  dplyr::left_join(patches_tib, turtles_tib)
+      agentdata <-  suppressMessages(dplyr::left_join(patches_tib, turtles_tib))
     }
 
   }
