### 2019-01-09 #262 @marcosci
# Detailed response to reviewers

Dear @marinapapa and @thomasp85,
thank you very much for your helpful and constructive feedback on our R package.
We agree with all the issues you mentioned and we revised our code and package documentation accordingly. The following sections will give you a detailed overview of our revisions, pointing to specific commits, where possible and links to the corresponding vignette/article sections of the corresponding package homepage.

## Reviewer comments by @thomasp85

### Documentation

> I feel the README could be a bit more detailed when it comes to describing what the package does. As it stands, you need to already be familiar with NetLogo to understand the need/use for the package. While a detailed description of NetLogo is not suited for a README, a small section before Installation would make the README more welcome to new users of actor based simulations in general. Further it would be good to link to the NetLogo online ressources to help users with researching further information. The README should also clearly state that you need to install NetLogo manually and how this is done. Omission of this indicates that NetLogo is installed as part of the package installation.

* We revised the README and added more specific information (see [README update commit 1](https://github.com/nldoc/nlrx/commit/0ccfb478a590b2516158f68e448c3ec774d3b360) and following, see also [home site](https://nldoc.github.io/nlrx/) of package homepage, built from README.Rmd)
* More specifically, we added more details on the purpose and usage of our package. We also added general information on NetLogo and links to NetLogo online resources.
* We added a Prerequirements section with installation guides and notes on NetLogo and Java

> The vignettes assumes the user is running on windows. While it may be impossible to create examples that run on all systems, as you have to point to the NetLogo installation it would be nice if the vignettes were structured in a way that made it easier to modify the code to your local system. This could e.g. be done by storing relevant paths in variables at the top of the vignette.

* We added a codeblock with default path definitions for windows and unix systems at the start of each vignette ([Get Started](https://nldoc.github.io/nlrx/articles/getstarted.html#step-1-create-a-nl-object), [Simdesign examples](https://nldoc.github.io/nlrx/articles/simdesign-examples.html#simdesign-examples-for-the-wolf-sheep-predation-model), [Capturing Spatial NetLogo Output](https://nldoc.github.io/nlrx/articles/articles/spatial-output.html#gathering-spatial-output-from-netlogo-model-simulations) and [Interface examples](https://nldoc.github.io/nlrx/articles/articles/interfacing-different-models.html))

> I would like the Simdesing examples vignette to include more descriptive text to carry the reader through what is happening, and giving them an impediment to understand the steps. As it is now, it is mainly a collections of snippets to cope-paste (or at least it appears like that with the lack of motivation and discussion)

* The revised [simdesign examples](https://nldoc.github.io/nlrx/articles/simdesign-examples.html) vignette contains more detailed information on the simdesign methods and how they can be used for model analysis. However, we only included very brief descriptions of the methods and linked to the relevant publications/R-packages for more detailed information on specific methods.

> Functions are in general well-documented. I think it would make sense to put .initClasses and simdesign-class documentation as internal so it doesn't show up in the overview, as the docs are rather lacking (and not that important).

* The class initialization function is now set as internal and does not show up in the functions overview table.

> All examples appear to be wrapped in \dontrun{}, which I'm pretty sure CRAN will take offense with. I can understand that running the actual simulations may be too heavy for an example, but a lot of the constructors and getters/setters are pretty benign and should be able to be run in examples

* The `\dontrun{}` wrappers have been removed where possible (see [commit](https://github.com/nldoc/nlrx/commit/eb937a546e8dffa17a5a6ff1b7e219d7cae75f1d))

### Functionality

> As discussed above, the assumption that NetLogo is already installed is not clearly communicated. The download_netlogo() function should be highlighted in the README and "Getting Started" vignette. Apart from this, the installation processes as expected, and the package works as expected. It is clear that you have to have substantial understanding of NetLogo prior to using this package, but as this is also the intended audience it is not really a problem (some additional hand-holding from the documentation would be nice as discussed above). The performance of the package is mainly a function of the performance of NetLogo, and it does not appear that the package adds any unnecessary computational overhead.

* As stated above, we completely revised the README and getstarted vignette accordingly and added more detailed information on NetLogo and the installation process. The download_netlogo() function is now also mentioned in our new [Prerequirements](https://nldoc.github.io/nlrx/articles/getstarted.html#prerequirements) section.

> The test coverage is generally good, and it is nice to see that they have made a strategy for testing functionality that can't be tested on CRAN

> The package title needs to be in titlecase and the description shouldn't start with "The nlrx package..." (CRAN requirements, not mine). Apart from that the packaging guidelines seems to be followed (again, with the excdeption of the documentation discussed above)

* We adjusted the title and the description accordingly.

## Reviewer comments by @marinapapa

> In general, I like the idea of the package and its functionalities. Some dependencies issues that were not stated made the review process a bit frustruating, but I think most of my points are very easy to tackle.

### Installation & Documentation
> 1. Installation worked in Windows, with the below problems:
>     * _R dependencies_:  R (>= 2.10) is mentioned in desciption, but two of the Imports packages (lhs and sf) need R >= 3.3. So nlrx could not be installed on a machine with older version.
>     * _Java_: the package is using a java virtual mashine so _Java_ needs to be installed on the local machine but is not mentioned anywhere. Installation is fine without it but of course the package will not function, the path set-up is needed in order to run.
> So it would be helpful if the information for the above points were stated in the installation guidelines/ README/ description.

* We adjusted the R-dependency information in the description file accordingly
* We revised our [README](https://nldoc.github.io/nlrx/index.html) and [get started](https://nldoc.github.io/nlrx/articles/getstarted.html) vignette and added a new section on [Prerequirements](https://nldoc.github.io/nlrx/index.html#prerequirements) that contains information on how to download and install NetLogo and instructions for Java installation.

> 2. A clearer statement of need is needed (as per rOpenSci guidelines), why the package is needed and should be preferred from just using Netlogo's behavioral space. Some of the text from the onboarding issue would fit nicely.

* Both, the [description](https://github.com/nldoc/nlrx/blob/master/DESCRIPTION) and the first sections of the [README](https://nldoc.github.io/nlrx/index.html) now contain more specific information on the purpose of the nlrx package. For instance, we now highlight the higher flexibility in designing experiments compared to NetLogo Behavior Space and the reproducible workflow.

> 3. _README_:
> * _Netlogo dependencies_: supported Netlogo versions are mentioned only in the _download_netlogo_ function. This and the above prerequests from (1) should be included in README.

* We added information on the supported NetLogo versions (>=5.3.1) to the new [Prerequirements](https://nldoc.github.io/nlrx/index.html#prerequirements) section of the README and get started vignette.

> * The first paragraph states what is the package doing in a very broad way. As mentioned above, a bit more emphasis can be given on stating the problems it is designed to solve and target audience.

* As mentioned above, the [introduction](https://nldoc.github.io/nlrx/index.html#nlrx) section of the README now contains more information on the purpose and main features of the package.

> * Example: this session could be renamed as a _Get started_ session and include the example. Also, starting with 'S4 classes' may be a bit too advanced for a student or a not-R-expert to understand, so I would restate the first paragraph in simpler words. The more techincal description is given at the DESCRIPTION file anyway, which I think is appropriate.

* As suggested, we renamed the example section to [Get started](https://nldoc.github.io/nlrx/index.html#get-started)
* We tried to explain the class concept in a less technical way. However, we think it is important to make the nested class concept of the nlrx package clear from the beginning.

> 4. _Help files_:
> Given that NetLogo is a program widely used for teaching, I think a bit more detailed help files chould be of great importance:
> 1. _nl contructor_: what are the experiment and simdesign should be explained more, or have a link to their help pages.

* We added information on experiment and simdesign classes to the [nl helpfile](https://nldoc.github.io/nlrx/reference/nl.html) including links to the corresponding documentation pages.

> 2. _experiment constructor_: should be more detailed of what information the experiment holds (not only with example code).

* We added more detailed information on the type of information that is entered into each experiment slot to the [experiment helpfile](https://nldoc.github.io/nlrx/reference/experiment.html).

> 3. _simdesign constructor_: design types should be listed in the description, not just stated as examples, maybe also with links to the separate functions descriptions.

* We added information on all supported simdesign types and helper functions to the [simdesign helpfile](https://nldoc.github.io/nlrx/reference/simdesign.html). We also added links to the corresponding helper function ducoumentation pages.

> 4. _analyze_nl_:
> * typo: fullstop missing in value session.

* revised

> * this function is based on the simdesign, but given that the latter is stored in the nl object, it is not so obvious for a new user. Maybe it should be emphasized a bit more in the help file and include links to the separate analyze functions.

* We added information on the different analyze_nl subfunctions to the [analyze_nl helpfile](https://nldoc.github.io/nlrx/reference/analyze_nl.html). We also added links to the corresponding help files of the analysis functions.

> * All the analyze_X functions have the same help file. If you think re-adding details of the analysis here (cause they exist in simdesign files) is not useful, you should at least add links to these helpfiles. Especially for:

* We added an overview list of analyze_X functions with brief explanations to the [analzye_nl helpfile](https://nldoc.github.io/nlrx/reference/analyze_nl.html). We also included links to the corresponding packages that are used for index calculations. We also added the brief explanation from this overview list to each analyze_X function helpfile.

> 5. _analyze_simple_: is an empty function, but nothing is stated in its help file.

* We removed all references to analyze_simple from the documentation as for now, no general useful postprocessing analysis can be done for just one simulation run. However, we kept the skeleton of the function in order to print a warning message if a nl object with a simple simdesign is parsed to the analyze_nl function.
* Additionally, we added a list of simdesign without postprocessing function to the [analyze_nl help file](
* https://nldoc.github.io/nlrx/reference/analyze_nl.html)

> 6. _simdesign_simple_: even though constant parameterization is stated, the fact that no analysis is available in this design should be mentioned.

* This is now stated in the [analyze_nl help file](https://nldoc.github.io/nlrx/reference/analyze_nl.html) and the analyze_simple help file.

> 7. _run_nl_one_ and _run_nl_all_ have the same title and description. A simple adjustment of 'Execute one NetLogo simulation' and 'Execute all NetLogo simulations' will do better.

* We adjusted the title and description of these functions accordingly ([run_nl_one](https://nldoc.github.io/nlrx/reference/run_nl_one.html), [run_nl_all](https://nldoc.github.io/nlrx/reference/run_nl_all.html))

> 8. _download_netlogo_: the saving path is stated as: 'Character string with the name where the downloaded file is saved.'. I think it should be replaced by 'path to folder' as in the rest of help me files with parth as input.

* We adjusted the description of the saving path accordingly ([download_netlogo](https://nldoc.github.io/nlrx/reference/download_netlogo.html))

> 9. _util_eval__: all functions help files are not very informative, in terms of what 'evaluate' means. However, the comments in the code of the functions are informative: 'check if there are any variables defined', so a simple copy paste should do.

* We adjusted the documentation of the evaluation functions accordingly (see [commit](https://github.com/nldoc/nlrx/commit/90c42adbec8e72f90cfe7818efc568bccdc565bf))

### Functions:

> **getstarted.Rmd**
> * library(nlrx) should be called after the installation code

* All code example section in the README and vignette files now contain a call of library(nlrx)

> * library(future) needs to be called before lines 100 and 102 can be implemented (otherwise operator %<-% and unquoted _multisession_ are not recognized)

* We removed the future operator from the basic examples in the README and [GetStarted](https://nldoc.github.io/nlrx/articles/getstarted.html) vignette for simplification. The usage of future in combination with nlrx is now explained within the [Advanced Configuration](https://nldoc.github.io/nlrx/articles/furthernotes.html#notes-on-parallelisation-and-the-future-concept) vignette, including a call to load the future package (library(future)).

> * If Java path is not connected and the user uses "multisession" plan gets an unhelpful error for a non-existing csv file in Temp. For different plans, they get error for Java path missing, which helps. I don't know if you can add something to that, maybe not use 'multisession' in the getstarted?. But anyway not that important.

* As stated above, we removed usage of future from the README and [GetStarted](https://nldoc.github.io/nlrx/articles/getstarted.html) vignette for simplification.

> * Attach an experiment: it should be mentioned if the user has to define all the model variables or the ones that are not defined will be assigned the default model's values.

* The fact that non-defined NetLogo variables will be set up constant with the default value from the NetLogo GUI is now stated within the [experiment help file](https://nldoc.github.io/nlrx/reference/experiment.html) and the [Advanced configuration](https://nldoc.github.io/nlrx/articles/furthernotes.html#notes-on-variables-and-constants-definitions) vignette.

> **nl@experiment**:
> * When assigning variables lists: if you add parameters X and Y as contants as well you get an error: 'Columns X, Y must have unique names', which doesn't make lots of sense. If you know the problem it is clear that it means that the variables should not be defined as contants, but the error message is not helping the user see whats wrong. A message can be added.

* We added an error message that is thrown when a simdesign is attached to a nl object with an experiment that has the same NetLogo parameter in the variables and constants list. The same message appears if such a nl object is commited to eval_variables_constants(). We also added a reference to this evaluation function to the [Advanced configuration](https://nldoc.github.io/nlrx/articles/furthernotes.html#notes-on-variables-and-constants-definitions) vignette.

> * If a parameter is wrong, after running the simulations, the results are going to be empty but no warning or other message will be shown before you try to assign in the results, and will get an error 'replacement has 1 row, data has 0'. It would be helful if there was a hint to the user to check the parameters in the experiment should be given.

* We added two warning messages to util_gather_results: (1) When using the systems default temp files directory, on linux systems it can happen that temporary files are deleted by system processes. If no csv file has been found for the current simulation, a warning message will appear, stating that the temporary files directory may need to be reassigned to another directory. (2) if the model output file is empty, a warning message will appear, stating that parameter definitions might be invalid or a runtime error might have occured.

> **spatial-output.Rmd**
> I couldn't get it to fully work. Two main problems:
> * line 86: _get_nl_spatial_ runs but prints a very bad message, a repetiting (a lot of times) 'no non-missing arguments to min; returning Infno non-missing arguments to max; returning -Inf', which I don't know where it comes from. With another model without patches information the message does not appear.

* We attached a .Rmd ([nlrx_spatial_debug.zip](https://github.com/ropensci/software-review/files/2741734/nlrx_spatial_debug.zip)) in which we tried to reproduce the warnings. With the current version of nlrx, it does not seem to produce these messages anymore. Can you confirm that?

> * line 137: for _results_spatial_tibble_, all metrics.turtles are returned NA. I think it comes from the _full_join_ of ddplyr, which only preserves the grouping of x (so pathces) and the not matching values are ignored and replaced with NA. So the turtles cannot be plotted afterwards, getting error 'object 'turtles_x' not found'

* The error was human induced - we just missed to update the vignette to the current version of nlrx. We fixed this now, `turtles_x` is not returned anymore. This is due to different ways agent coordinates can be returned, either x or pc. nlrx now returns this as column name.

> And minor:
> * _library(ggplot2)_ needs to be called earlier, before first plot of spatial output.

* We added a call to library(ggplot2) at the start of the [Spatial Output code example](https://nldoc.github.io/nlrx/articles/articles/spatial-output.html#gathering-spatial-output-from-netlogo-model-simulations).

> * _gganimate_: also needs a renderer installed. I would mention or add the command for installing 'gifski' package (the default of gganimate), and also the 'png' package installed (which is needed by gganimate). Also, maybe add hashed-out the direct gganimate installation command (instead of just the link).

* We added a comment on the gifski dependency for gganimate at the start of the [Spatial Output code example](https://nldoc.github.io/nlrx/articles/articles/spatial-output.html#gathering-spatial-output-from-netlogo-model-simulations).

> **analyze_simple**
> Renaming to _no_analysis_ may be helpful (a 'simple' analysis is still an analysis). So far if the user runs _analyze_nl_ with simple design they get NULL, which can be confusing. A message can be added to be printed in _analuze_simple_ in order to inform the user that 'no analysis can take place if a simple design is chosen'.

* We added a print message to analyze_simple stating that no analysis is implemented yet for simdesign_simple.

> **write_simoutput**
[…]
