### 2019-01-10 #262 @annakrystalli
Thank you @marcosci & @nldoc for the detailed response and improvements!

@thomasp85 & @marinapapa , over to you. Please feedback on whether the responses have adequately addressed the issues you've raised. Feel free to reach out for any clarification.
