### 2019-01-13 #262 @marinapapa
Dear Jan and Marco,

I am sorry for not responding to your initial message about **get_nl_spatial** . It is indeed running smoothly now so no need for anything else :)

The new **documentation** is great and all my initial issues are adequately addressed.

I also like very very much the **interfacing-different-models** addition! Some comments:
- A typo: on the fire model example (line 130) you are inputting the diffusion model again.
- Ants example:
1. scale_color_gradient_tableau (line 267) needs library(ggthemes)
2. animate(line 275) throws error: 'time data must be the same class in all layers'

- Waves example:  line 322 run_nl_all gives me a message
"you must specify either --setup-file or --experiment (or both)
'Physics' is not recognized as an internal or external command,
operable program or batch file. " followed by an error:
"Error in util_gather_results(nl, outfile, seed, siminputrow) : Temporary output file C:\..\Temp\RtmpCIfUkx\nlrx7380_125bc5e046c5f.csvnot found. On unix systems this can happen if the default system temp folder is used. Try reassigning the default temp folder for this R session (unixtools package)."

I didn't try much to solve the errors, maybe it is something very straight forward. Everything else works nicely :)
