### 2019-01-14 0f69528
fix ants example
