### 2019-01-14 #262 @marcosci
Hi Marina,

thanks again for the comments!
We fixed all the issues you mentioned in the vignette - the wave example was mainly due to OS differences between Jan and myself.

Cheers,
Marco

PS: If you ever use gganimate for real flocking data or more complex models, please let me know :) would love to see that!
