### 2019-01-14 #262 @thomasp85
Hi Jan and Marco

I can only echo Marina in saying that all raised issues seems to have been solved. You have gone all-in on providing good documentation and guidance which warms my heart :-)

I do not have any additional comments to the package as it appears now.

best
Thomas
