### 2019-01-15 54f37e6
include ropensci footer and change adress of covr, travis and appveyor
