### 2019-01-16 #262 @marcosci
@annakrystalli  that was just a remnant of the beginning of the package - I updated the badge, as we consider it quite stable now :+1:
