### 2019-01-22 #262 @stefaniebutland
Hello @nldoc and @marcosci and congratulations on passing peer review! Another one for you Marco.

Are you interested in writing a blog post or tech note?  This link will give you many examples of blog posts by authors of peer-reviewed packages so you can get an idea of the style and length you prefer: https://ropensci.org/tags/software-peer-review/.

Technotes are here: https://ropensci.org/technotes/.

Here are some technical and editorial guidelines for contributing a post: https://github.com/ropensci/roweb2#contributing-a-blog-post.

Happy to answer any questions.
